### 2024-02-23 c908057
Re-typed JOSS paper metadata while sticking strictly to template to avoid errors
