+++ docs/JOSS_Paper/paper.md
@@ -5,19 +5,58 @@ tags:
   - Radiative Transfer
   - Discrete Ordinates Method
   - Atmospheric Science
-  - Climate Science
+  - Climate Models
   - DISORT
 authors:
   - name: Dion J. X. Ho
-	orcid: 0009-0000-5829-5081
-    affiliation: 1
+    orcid: 0009-0000-5829-5081
+    equal-contrib: true
+    affiliation: "1"
 affiliations:
  - name: Columbia University, Department of Applied Physics and Applied Mathematics, USA
    index: 1
-date: 11 February 2023
-bibliography: PythonicDISORT.bib
-
+date: 11 February 2024
+bibliography: paper.bib
 ---
+
+<!---
+title: 'Gala: A Python package for galactic dynamics'
+tags:
+  - Python
+  - astronomy
+  - dynamics
+  - galactic dynamics
+  - milky way
+authors:
+  - name: Adrian M. Price-Whelan
+    orcid: 0000-0000-0000-0000
+    equal-contrib: true
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Author Without ORCID
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+  - name: Author with no affiliation
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: 3
+  - given-names: Ludwig
+    dropping-particle: van
+    surname: Beethoven
+    affiliation: 3
+affiliations:
+ - name: Lyman Spitzer, Jr. Fellow, Princeton University, USA
+   index: 1
+ - name: Institution Name, Country
+   index: 2
+ - name: Independent Researcher, Country
+   index: 3
+date: 13 August 2017
+bibliography: paper.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-journal: Astrophysical Journal <- The name of the AAS journal.
+--->
 
 # Summary
 The Radiative Transfer Equation (RTE) models the processes of absorption, scattering and emission 
