### 2024-02-23 #6399 @dhjx1996
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/8022028259?check_suite_focus=true).

I have pushed changes to address the error
