### 2024-02-23 #6399 @dhjx1996
Suggested reviewers: [arjunsavel](https://reviewers.joss.theoj.org/reviewers/1410), [Peter Scicluna](https://reviewers.joss.theoj.org/reviewers/3041), [simonrp84](https://reviewers.joss.theoj.org/reviewers/1941) because Radiative Transfer is in their topic area.
