### 2024-03-04 #6442 @editorialbot
[REVIEW]: PythonicDISORT: A Python reimplementation of the Discrete Ordinate Radiative Transfer package DISORT
**Submitting author:** @dhjx1996 (<a href="http://orcid.org/0009-0000-5829-5081">Jia Xu Dion Ho</a>)
**Repository:** https://github.com/LDEO-CREW/Pythonic-DISORT
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.3
