### 2024-03-04 #6442 @phibeck
@dhjx1996 as reported above by editorialbot, there are two references with invalid DOIs. When you have a moment, please fix these, thanks
