### 2024-03-04 11733b0
Corrected DOIs for JOSS paper and made slight changes to README
