+++ docs/JOSS_Paper/paper.bib
@@ -40,7 +40,7 @@ number = {5},
 pages = {443-451},
 year = {1998},
 issn = {0098-3004},
-doi = {https://doi.org/10.1016/S0098-3004(97)00130-1},
+doi = {10.1016/S0098-3004(97)00130-1},
 url = {https://www.sciencedirect.com/science/article/pii/S0098300497001301},
 author = {Jeffrey R Key and Axel J Schweiger},
 keywords = {Radiative transfer models, Atmospheric processes, Radiation budget, Satellites},
@@ -56,7 +56,7 @@ abstract = {Two tools for the solution of radiative transfer problems are presen
       address = "Boston MA, USA",
       volume = "79",
       number = "10",
-      doi = "https://doi.org/10.1175/1520-0477(1998)079<2101:SARATS>2.0.CO;2",
+      doi = "10.1175/1520-0477(1998)079<2101:SARATS>2.0.CO;2",
       pages=      "2101 - 2114",
       url = "https://journals.ametsoc.org/view/journals/bams/79/10/1520-0477_1998_079_2101_sarats_2_0_co_2.xml"
 }
