### 2024-03-04 #6442 @dhjx1996
> @dhjx1996 as reported above by editorialbot, there are two references with invalid DOIs. When you have a moment, please fix these, thanks

I have fixed these and pushed the changes.
