### 2024-03-12 9fd7732
Computation of eigenpairs is now completely vectorized (no for loops)
