+++ src/PythonicDISORT/_solve_for_coefs.py
@@ -1,0 +1,234 @@
+from PythonicDISORT.subroutines import _mathscr_v
+import numpy as np
+import scipy as sc
+from math import pi
+
+
+def _solve_for_coefs(
+    NLoops,
+    G_collect,
+    K_collect,
+    B_collect,
+    G_inv_collect_0,
+    tau_arr,
+    scaled_tau_arr_with_0,
+    mu_arr_pos, mu_arr_pos_times_W, mu_arr,
+    N, NQuad,
+    NLayers, NBDRF,
+    BDRF_Fourier_modes,
+    mu0, I0,
+    b_pos, b_neg,
+    scalar_b_pos, scalar_b_neg,
+    s_poly_coeffs,
+    Nscoeffs,
+    use_sparse_NLayers,
+):
+    """This function is wrapped and looped by the `_assemble_results` function.
+    It has many seemingly redundant arguments to maximize precomputation in the `pydisort` function.
+    See the Jupyter Notebook, especially section 3, for documentation, explanation and derivation.
+    The labels in this file reference labels in the Jupyter Notebook, especially sections 3 and 4.
+
+    """
+    ################################## Solve for coefficients of homogeneous solution ##########################################
+    GC_collect = np.empty((NLoops, NLayers, NQuad, NQuad))
+    
+    for m in range(NLoops):
+        G_collect_m = G_collect[m, :, :, :]
+        K_collect_m = K_collect[m, :, :]
+        if I0 > 0:
+            B_collect_m = B_collect[m, :, :]
+            
+        # Generate mathscr_D and mathscr_X (BDRF terms)
+        # --------------------------------------------------------------------------------------------------------------------------
+        if m < NBDRF:
+            mathscr_D_neg = (1 + (m == 0) * 1) * BDRF_Fourier_modes[m](mu_arr_pos, -mu_arr_pos)
+            R = mathscr_D_neg * mu_arr_pos_times_W[None, :]
+
+            if I0 > 0:
+                mathscr_X_pos = (mu0 * I0 / pi) * BDRF_Fourier_modes[m](
+                    mu_arr_pos, -np.array([mu0])
+                )[:, 0]
+        # --------------------------------------------------------------------------------------------------------------------------
+    
+        # Assemble RHS
+        # --------------------------------------------------------------------------------------------------------------------------
+        # Ensure the BCs are of the correct shape
+        if scalar_b_pos:
+            if m == 0:
+                b_pos_m = np.full(N, b_pos)
+            else:
+                b_pos_m = np.zeros(N)
+        else:
+            b_pos_m = b_pos[:, m]
+        if scalar_b_neg:
+            if m == 0:
+                b_neg_m = np.full(N, b_neg)
+            else:
+                b_neg_m = np.zeros(N)
+        else:
+            b_neg_m = b_neg[:, m]
+        
+        # _mathscr_v_contribution
+        if Nscoeffs > 0 and m == 0:
+            _mathscr_v_contribution_top = -_mathscr_v(
+                        np.array([0]), np.array([0]),
+                        s_poly_coeffs[0:1, :],
+                        Nscoeffs,
+                        G_collect_m[0:1, N:, :],
+                        K_collect_m[0:1, :],
+                        G_inv_collect_0[0:1, :, :],
+                        mu_arr,
+                    ).flatten()
+        
+            _mathscr_v_contribution_middle = np.array([])
+            if NLayers > 1:
+                indices = np.arange(NLayers - 1)
+                _mathscr_v_contribution_middle = (
+                    _mathscr_v(
+                        tau_arr[:-1],
+                        indices,
[…]
