+++ src/PythonicDISORT/_diagonalize.py
@@ -1,0 +1,179 @@
+import numpy as np
+import scipy as sc
+from math import pi
+
+
+def _diagonalize(
+    NLoops,
+    scaled_omega_arr,
+    mu_arr_pos, mu_arr,
+    M_inv, W,
+    N, NQuad, NLeg,
+    NLayers,
+    weighted_scaled_Leg_coeffs,
+    mu0, I0,
+    Nscoeffs,
+):
+    """This function is wrapped and looped by the `_assemble_results` function.
+    It has many seemingly redundant arguments to maximize precomputation in the `pydisort` function.
+    See the Jupyter Notebook, especially section 3, for documentation, explanation and derivation.
+    The labels in this file reference labels in the Jupyter Notebook, especially sections 3 and 4.
+
+    """
+    ells_all = np.arange(NLeg)
+    ind = 0
+    G_collect = np.empty((NLoops * NLayers, NQuad, NQuad))
+    K_collect = np.empty((NLoops * NLayers, NQuad))
+    if I0 > 0:
+        B_collect = np.zeros((NLoops * NLayers, NQuad))
+    if Nscoeffs > 0:
+        G_inv_collect_0 = np.empty((NLayers, NQuad, NQuad))
+    alpha_list = []
+    beta_list = []
+    X_tilde_list=[]
+    if_indices = []
+
+    # Loop over NLoops Fourier modes
+    # --------------------------------------------------------------------------------------------------------------------------  
+    for m in range(NLoops): 
+        # Setup
+        # --------------------------------------------------------------------------------------------------------------------------
+        ells = ells_all[m:]
+        degree_tile = np.tile(ells, (N, 1)).T
+        fac = sc.special.poch(ells + m + 1, -2 * m)
+        signs = np.ones(NLeg - m)
+        signs[1::2] = -1
+
+        asso_leg_term_pos = sc.special.lpmv(m, degree_tile, mu_arr_pos)
+        asso_leg_term_neg = asso_leg_term_pos * signs[:, None]
+        asso_leg_term_mu0 = sc.special.lpmv(m, ells, -mu0)
+        # --------------------------------------------------------------------------------------------------------------------------
+
+        # Loop over NLayers atmospheric layers
+        # --------------------------------------------------------------------------------------------------------------------------          
+        for l in range(NLayers):
+            # More setup
+            weighted_asso_Leg_coeffs_l = weighted_scaled_Leg_coeffs[l, :][ells] * fac
+            scaled_omega_l = scaled_omega_arr[l]
+            
+            # We take precautions against overflow and underflow (this shouldn't happen though)
+            if np.any(weighted_asso_Leg_coeffs_l > 0) and np.all(
+                np.isfinite(asso_leg_term_pos)
+            ):  
+                # Generate mathscr_D and mathscr_X (BDRF terms)
+                # --------------------------------------------------------------------------------------------------------------------------
+                D_temp = weighted_asso_Leg_coeffs_l[None, :] * asso_leg_term_pos.T
+                D_pos = (scaled_omega_l / 2) * D_temp @ asso_leg_term_pos
+                D_neg = (scaled_omega_l / 2) * D_temp @ asso_leg_term_neg
+
+                if I0 > 0:
+                    X_temp = (
+                        (scaled_omega_l * I0 * (2 - (m == 0)) / (4 * pi))
+                        * weighted_asso_Leg_coeffs_l
+                        * asso_leg_term_mu0
+                    )
+                    X_pos = X_temp @ asso_leg_term_pos
+                    X_neg = X_temp @ asso_leg_term_neg
+                # --------------------------------------------------------------------------------------------------------------------------
+
+                # Assemble the coefficient matrix and additional terms
[…]

[…]
