+++ src/PythonicDISORT/pydisort.py
@@ -1,5 +1,5 @@
 from PythonicDISORT import subroutines
-from PythonicDISORT._assemble_results import _assemble_results
+from PythonicDISORT._assemble_solution_functions import _assemble_solution_functions
 import scipy as sc
 from math import pi
 from numpy.polynomial.legendre import Legendre
@@ -126,6 +126,9 @@ def pydisort(
     thickness_arr = np.concatenate([[tau_arr[0]], np.diff(tau_arr)])
     NLeg_all = np.shape(Leg_coeffs_all)[1]
     N = NQuad // 2
+    beam_source_bool = I0 > 0
+    iso_source_bool = Nscoeffs > 0
+    multilayer_bool = NLayers > 1
     # --------------------------------------------------------------------------------------------------------------------------
 
     # Input checks
@@ -148,7 +151,7 @@ def pydisort(
     assert len(omega_arr) == NLayers
     if len(f_arr) != 1 or f_arr[0] != 0:
         assert len(f_arr) == NLayers
-    if Nscoeffs > 0:
+    if iso_source_bool:
         assert np.shape(s_poly_coeffs)[0] == NLayers
     # Conditions on the number of quadrature angles (NQuad), Legendre coefficients (NLeg) and loops (NLoops)
     assert NQuad >= 2
@@ -159,7 +162,7 @@ def pydisort(
     assert NQuad >= NLeg
     # We require principal angles and a downward incident beam
     assert I0 >= 0
-    if I0 > 0:
+    if beam_source_bool:
         assert 0 < mu0 and mu0 <= 1
         assert 0 <= phi0 and phi0 < 2 * pi
     # Ensure that the BC inputs are of the correct shape
@@ -182,7 +185,7 @@ def pydisort(
     NBDRF = len(BDRF_Fourier_modes)
     weighted_Leg_coeffs_all = (2 * np.arange(NLeg_all) + 1) * Leg_coeffs_all
     Leg_coeffs = Leg_coeffs_all[:, :NLeg]
-    if (scalar_b_pos and b_pos == 0) and (scalar_b_neg and b_neg == 0) and Nscoeffs == 0 and I0 > 0:
+    if (scalar_b_pos and b_pos == 0) and (scalar_b_neg and b_neg == 0) and not iso_source_bool and beam_source_bool:
         I0_orig = I0
         I0 = 1
     else:
@@ -220,12 +223,12 @@ def pydisort(
         scaled_omega_arr = omega_arr
     # --------------------------------------------------------------------------------------------------------------------------
     
-    if NT_cor and not only_flux and I0 > 0 and np.any(f_arr > 0) and NLeg < NLeg_all:
+    if NT_cor and not only_flux and beam_source_bool and np.any(f_arr > 0) and NLeg < NLeg_all:
         
         ############################### Perform NT corrections on the intensity but not the flux ###################################
         
         # Delta-M scaled solution; no further corrections to the flux
-        flux_up, flux_down, u0, u_star = _assemble_results(
+        flux_up, flux_down, u0, u_star = _assemble_solution_functions(
             scaled_omega_arr,
             tau_arr,
             scaled_tau_arr_with_0,
@@ -233,13 +236,16 @@ def pydisort(
             M_inv, W,
             N, NQuad, NLeg, NLoops,
             NLayers, NBDRF,
+            multilayer_bool,
             weighted_scaled_Leg_coeffs,
             BDRF_Fourier_modes,
             mu0, I0, I0_orig, phi0,
+            beam_source_bool,
             b_pos, b_neg,
             scalar_b_pos, scalar_b_neg,
             s_poly_coeffs,
             Nscoeffs,
+            iso_source_bool,
             scale_tau,
             only_flux,
             use_sparse_NLayers,
[…]
