+++ src/PythonicDISORT/_diagonalize.py
@@ -12,27 +12,36 @@ def _diagonalize(
     NLayers,
     weighted_scaled_Leg_coeffs,
     mu0, I0,
+    beam_source_bool,
     Nscoeffs,
+    iso_source_bool,
 ):
-    """This function is wrapped by the `_assemble_results` function.
+    """This function is wrapped by the `_assemble_solution_functions` function.
     It has many seemingly redundant arguments to maximize precomputation in the `pydisort` function.
     See the Jupyter Notebook, especially section 3, for documentation, explanation and derivation.
     The labels in this file reference labels in the Jupyter Notebook, especially sections 3 and 4.
 
     """
+    ############################### Assemble system and diagonalize coefficient matrix #########################################
+    
+    # Initialization
+    # --------------------------------------------------------------------------------------------------------------------------
     ells_all = np.arange(NLeg)
     ind = 0
+    
     G_collect = np.empty((NLoops * NLayers, NQuad, NQuad))
     K_collect = np.empty((NLoops * NLayers, NQuad))
-    if I0 > 0:
-        B_collect = np.zeros((NLoops * NLayers, NQuad))
-    if Nscoeffs > 0:
-        G_inv_collect_0 = np.empty((NLayers, NQuad, NQuad))
-    alpha_list = []
-    beta_list = []
-    X_tilde_list=[]
+    alpha_arr = np.empty((NLoops * NLayers, N, N))
+    beta_arr = np.empty((NLoops * NLayers, N, N))
+    X_tilde_arr = np.empty((NLoops * NLayers, NQuad))
     no_shortcut_indices = []
     no_shortcut_indices_0 = []
+    
+    if beam_source_bool:
+        B_collect = np.zeros((NLoops * NLayers, NQuad))
+    if iso_source_bool:
+        G_inv_collect_0 = np.empty((NLayers, NQuad, NQuad))
+    # --------------------------------------------------------------------------------------------------------------------------
 
     # Loop over NLoops Fourier modes
     # These can easily be parallelized, but the speed-up is unlikely to be worth the overhead
@@ -40,6 +49,7 @@ def _diagonalize(
     for m in range(NLoops): 
         # Setup
         # --------------------------------------------------------------------------------------------------------------------------
+        m_equals_0_bool = (m == 0)
         ells = ells_all[m:]
         degree_tile = np.tile(ells, (N, 1)).T
         fac = sc.special.poch(ells + m + 1, -2 * m)
@@ -68,7 +78,7 @@ def _diagonalize(
                 D_pos = (scaled_omega_l / 2) * D_temp @ asso_leg_term_pos
                 D_neg = (scaled_omega_l / 2) * D_temp @ asso_leg_term_neg
 
-                if I0 > 0:
+                if beam_source_bool:
                     X_temp = (
                         (scaled_omega_l * I0 * (2 - (m == 0)) / (4 * pi))
                         * weighted_asso_Leg_coeffs_l
@@ -82,14 +92,14 @@ def _diagonalize(
                 # --------------------------------------------------------------------------------------------------------------------------
                 alpha = M_inv[:, None] * (D_pos * W[None, :] - np.eye(N))
                 beta = M_inv[:, None] * D_neg * W[None, :]
-                if I0 > 0:
-                    X_tilde_list.append(np.concatenate([-M_inv * X_pos, M_inv * X_neg]))
+                if beam_source_bool:
+                    X_tilde_arr[ind, :] = np.concatenate([-M_inv * X_pos, M_inv * X_neg])
                 # --------------------------------------------------------------------------------------------------------------------------
                 
+                alpha_arr[ind, :, :] = alpha
[…]

[…]
