### 2024-03-13 f2f4594
Made changes to README and other docs. Ready to release version 0.5.0.
