+++ PythonicDISORT_optional_dependencies.txt
@@ -1,5 +1,4 @@
 # Optional dependencies for the PythonicDISORT package
 # Install through "pip install -r PythonicDISORT_optional_dependencies.txt"
 
-joblib
 pytest
