+++ .github/workflows/continuous_distribution.yml
@@ -54,5 +54,5 @@ jobs:
         uses: pypa/gh-action-pypi-publish@v1.6.4
         with:
           user: __token__
-          password: ${{ secrets.PYPI_TOKEN }}
+          password: ${{ secrets.UPLOAD_TO_PYPI }}
           verbose: true
