### 2024-03-15 779c891
Modified sparse framework. Commit before completing changes
