+++ src/PythonicDISORT/_solve_for_coefs.py
@@ -1,4 +1,5 @@
 from PythonicDISORT.subroutines import _mathscr_v
+from PythonicDISORT.subroutines import _nd_slice_to_indexes
 
 import numpy as np
 import scipy as sc
@@ -16,15 +17,15 @@ def _solve_for_coefs(
     mu_arr_pos, mu_arr_pos_times_W, mu_arr,
     N, NQuad,
     NLayers, NBDRF,
-    multilayer_bool,
+    atmos_is_multilayered,
     BDRF_Fourier_modes,
     mu0, I0,
-    beam_source_bool,
+    there_is_beam_source,
     b_pos, b_neg,
-    scalar_b_pos, scalar_b_neg,
+    b_pos_is_scalar, b_neg_is_scalar,
     s_poly_coeffs,
     Nscoeffs,
-    iso_source_bool,
+    there_is_iso_source,
     use_sparse_NLayers,
 ):
     """This function is wrapped by the `_assemble_solution_functions` function.
@@ -36,25 +37,25 @@ def _solve_for_coefs(
     ################################## Solve for coefficients of homogeneous solution ##########################################
     
     GC_collect = np.empty((NLoops, NLayers, NQuad, NQuad))
-    use_sparse_bool = NLayers >= use_sparse_NLayers
+    use_sparse_framework = NLayers >= use_sparse_NLayers
     
     # The following loops can easily be parallelized, but the speed-up is unlikely to be worth the overhead
     for m in range(NLoops):
-        m_equals_0_bool = (m == 0)
+        m_equals_0 = (m == 0)
         BDRF_bool = m < NBDRF
         
         G_collect_m = G_collect[m, :, :, :]
         K_collect_m = K_collect[m, :, :]
-        if beam_source_bool:
+        if there_is_beam_source:
             B_collect_m = B_collect[m, :, :]
             
         # Generate mathscr_D and mathscr_X (BDRF terms)
         # --------------------------------------------------------------------------------------------------------------------------
         if BDRF_bool:
-            mathscr_D_neg = (1 + m_equals_0_bool * 1) * BDRF_Fourier_modes[m](mu_arr_pos, -mu_arr_pos)
+            mathscr_D_neg = (1 + m_equals_0 * 1) * BDRF_Fourier_modes[m](mu_arr_pos, -mu_arr_pos)
             R = mathscr_D_neg * mu_arr_pos_times_W[None, :]
 
-            if beam_source_bool:
+            if there_is_beam_source:
                 mathscr_X_pos = (mu0 * I0 / pi) * BDRF_Fourier_modes[m](
                     mu_arr_pos, -np.array([mu0])
                 )[:, 0]
@@ -63,15 +64,15 @@ def _solve_for_coefs(
         # Assemble RHS
         # --------------------------------------------------------------------------------------------------------------------------
         # Ensure the BCs are of the correct shape
-        if scalar_b_pos:
-            if m_equals_0_bool:
+        if b_pos_is_scalar:
+            if m_equals_0:
                 b_pos_m = np.full(N, b_pos)
             else:
                 b_pos_m = np.zeros(N)
         else:
             b_pos_m = b_pos[:, m]
-        if scalar_b_neg:
-            if m_equals_0_bool:
+        if b_neg_is_scalar:
+            if m_equals_0:
                 b_neg_m = np.full(N, b_neg)
             else:
                 b_neg_m = np.zeros(N)
@@ -79,7 +80,7 @@ def _solve_for_coefs(
             b_neg_m = b_neg[:, m]
         
         # _mathscr_v_contribution
-        if iso_source_bool and m_equals_0_bool:
+        if there_is_iso_source and m_equals_0:
[…]
