+++ src/PythonicDISORT/pydisort.py
@@ -23,7 +23,7 @@ def pydisort(
     BDRF_Fourier_modes=[],
     s_poly_coeffs=np.array([[]]),
     use_sparse_NLayers=13,
-    _autograd_bool = False,
+    _is_compatible_with_autograd = False,
 ):
     """Solves the 1D RTE for the fluxes, and optionally intensity,
     of a multi-layer atmosphere with the specified optical properties, boundary conditions
@@ -96,7 +96,7 @@ def pydisort(
         the Cauchy / Fourier convergence evaluation (type: float) for the last Fourier term.
 
     """
-    if _autograd_bool:
+    if _is_compatible_with_autograd:
         import autograd.numpy as np
     else:
         import numpy as np
@@ -253,7 +253,7 @@ def pydisort(
             scale_tau,
             only_flux,
             use_sparse_NLayers,
-            _autograd_bool,
+            _is_compatible_with_autograd,
         )
         
         # TMS correction
@@ -318,88 +318,162 @@ def pydisort(
                 )
             )
             
-            # Contribution from other layers
+            ## Contribution from other layers
             # TODO: Despite being vectorized and otherwise optimized, this block of code is slower than desired
             # and is the bottleneck for the TMS correction
             # --------------------------------------------------------------------------------------------------------------------------
             if atmos_is_multilayered:
-
                 layers_arr = np.arange(NLayers)[:, None]
                 pos_contribution_mask = (l[None, :] < layers_arr).flatten()
                 neg_contribution_mask = (l[None, :] > layers_arr).flatten()
-
+                
                 layers_arr_repeat = np.repeat(layers_arr, Ntau)
                 scaled_tau_tile = np.tile(scaled_tau, NLayers)
                 scaled_tau_arr_with_0_repeat = np.repeat(scaled_tau_arr_with_0, Ntau)
-
-                if np.any(pos_contribution_mask):
-                    contribution_from_other_layers_pos = np.zeros((N, NLayers * Ntau, Nphi))
-                    scaled_tau_l_tile_pos = scaled_tau_tile[pos_contribution_mask]
-                    mathscr_B_l_repeat_posmasked = mathscr_B[:N, layers_arr_repeat[pos_contribution_mask], :]
-                    scaled_tau_arr_with_0_l_tile_posmasked = scaled_tau_arr_with_0_repeat[:-Ntau][
-                        pos_contribution_mask
-                    ]
-                    scaled_tau_arr_with_0_lp1_tile_posmasked = scaled_tau_arr_with_0_repeat[Ntau:][
-                        pos_contribution_mask
-                    ]
                     
-                    contribution_from_other_layers_pos[:, pos_contribution_mask, :] = (
-                        mathscr_B_l_repeat_posmasked
-                        * (
-                            np.exp(
-                                (scaled_tau_l_tile_pos - scaled_tau_arr_with_0_l_tile_posmasked)
-                                / mu_arr_pos[:, None]
-                                - scaled_tau_arr_with_0_l_tile_posmasked / mu0
-                            )
-                            - np.exp(
-                                (scaled_tau_l_tile_pos - scaled_tau_arr_with_0_lp1_tile_posmasked)
-                                / mu_arr_pos[:, None]
-                                - scaled_tau_arr_with_0_lp1_tile_posmasked / mu0
-                            )
-                        )[:, :, None]
-                    )
-                    contribution_from_other_layers_pos = np.sum(
-                        contribution_from_other_layers_pos.reshape((N, NLayers, Ntau, Nphi)), axis=1
-                    )
-                if np.any(neg_contribution_mask):
[…]
