### 2024-03-16 539f24f
Made substantial optimizations. Ready for release of v0.6.0.
