### 2024-03-16 f6749da
Replaced `flatten` with `ravel`
