### 2024-03-17 25ad246
Release v0.6.1 ready. We fixed a bug and reverted a change from the previous release (`np.add.at` is not fast enough).
