+++ src/PythonicDISORT/pydisort.py
@@ -321,159 +321,100 @@ def pydisort(
             )
             
             ## Contribution from other layers
-            # TODO: Despite being vectorized and otherwise optimized, this block of code is slower than desired
-            # and is the bottleneck for the TMS correction
+            # TODO: Despite being vectorized and despite many attempts at further optimization,
+            # this block of code is slower than desired and is the bottleneck for the TMS correction.
             # --------------------------------------------------------------------------------------------------------------------------
             if atmos_is_multilayered:
                 layers_arr = np.arange(NLayers)[:, None]
                 pos_contribution_mask = (l[None, :] < layers_arr).ravel()
                 neg_contribution_mask = (l[None, :] > layers_arr).ravel()
-                
+
                 layers_arr_repeat = np.repeat(layers_arr, Ntau)
                 scaled_tau_tile = np.tile(scaled_tau, NLayers)
                 scaled_tau_arr_with_0_repeat = np.repeat(scaled_tau_arr_with_0, Ntau)
-                    
-                if True:
-                    if np.any(pos_contribution_mask):
-                        contribution_from_other_layers_pos = np.zeros((N, NLayers * Ntau, Nphi))
-                        scaled_tau_l_tile_pos = scaled_tau_tile[pos_contribution_mask]
-                        mathscr_B_l_repeat_posmasked = mathscr_B[:N, layers_arr_repeat[pos_contribution_mask], :]
-                        scaled_tau_arr_with_0_l_tile_posmasked = scaled_tau_arr_with_0_repeat[:-Ntau][
-                            pos_contribution_mask
-                        ]
-                        scaled_tau_arr_with_0_lp1_tile_posmasked = scaled_tau_arr_with_0_repeat[Ntau:][
-                            pos_contribution_mask
-                        ]
-                        
-                        contribution_from_other_layers_pos[:, pos_contribution_mask, :] = (
-                            mathscr_B_l_repeat_posmasked
-                            * (
-                                np.exp(
-                                    (scaled_tau_l_tile_pos - scaled_tau_arr_with_0_l_tile_posmasked)
-                                    / mu_arr_pos[:, None]
-                                    - scaled_tau_arr_with_0_l_tile_posmasked / mu0
-                                )
-                                - np.exp(
-                                    (scaled_tau_l_tile_pos - scaled_tau_arr_with_0_lp1_tile_posmasked)
-                                    / mu_arr_pos[:, None]
-                                    - scaled_tau_arr_with_0_lp1_tile_posmasked / mu0
-                                )
-                            )[:, :, None]
-                        )
-                        contribution_from_other_layers_pos = np.sum(
-                            contribution_from_other_layers_pos.reshape((N, NLayers, Ntau, Nphi)), axis=1
-                        )
-                    if np.any(neg_contribution_mask):
-                        contribution_from_other_layers_neg = np.zeros((N, NLayers * Ntau, Nphi))
-                        scaled_tau_l_tile_neg = scaled_tau_tile[neg_contribution_mask]
-                        mathscr_B_l_repeat_negmasked = mathscr_B[N:, layers_arr_repeat[neg_contribution_mask], :]
-                        scaled_tau_arr_with_0_l_tile_negmasked = scaled_tau_arr_with_0_repeat[:-Ntau][
-                            neg_contribution_mask
-                        ]
-                        scaled_tau_arr_with_0_lp1_tile_negmasked = scaled_tau_arr_with_0_repeat[Ntau:][
-                            neg_contribution_mask
-                        ]
-                        
-                        contribution_from_other_layers_neg[:, neg_contribution_mask, :] = (
-                            mathscr_B_l_repeat_negmasked
-                            * (
-                                np.exp(
[…]
