### 2024-03-19 a58eb1a
Renamed "NLoops" to "NFourier". Ready to release v0.6.2.
