### 2024-04-15 #6442 @arjunsavel
Thank you for your patience. I’ve finished my review of the PythonicDISORT package. It’s an excellent piece of work that has clear utility in both pedagogical and research contexts.

My comments are below. Provided they’re addressed, I recommend this project for acceptance.

Miscellaneous
- This seems like quite a large GitHub repo. I believe 695 MB?
    - Could some of the size be cut down by trimming notebook histories? And evaluate the notebooks on the fly?
- The benefit of downloading the optional dependencies is not immediately apparent.
- No community guidelines. Maybe make a contributing.md?
- Really impressive set of notebooks!
- Thoughts on including polarization?

Code edits
- Could you please add function documentation in _one_Fourier_mode? For each argument?
- Same with _loop_and_assemble_results. Even though these aren’t ever accessed directly, documenting these makes for easier contributor / developer experience.
- pydisort is a bit of a long function and could likely be modularized. Is it implemented this way to be more one-to-one with the original fortran? Either way, this isn’t something that gets in the way of acceptance — I’m mostly just curious.
- Some code could maybe be vectorized — e.g. in src/PythonicDISORT/_diagonalize.py. Again, not required for acceptance.
- Could you add a more descriptive docstring to _assemble_solution_functions?

Paper edits
- Can we get definitions of the terms in the equations?
- Are omega and D functions of tau?
- You mention computational reasons…are those temporary? Is speed optimization a future goal, e.g., with numba, or Jax, or python? Or even joblib for parallelization? If you end up using Jax, you could replace your autograd usage with jax.
- Is “comprehensive documentation” meant to be hyperlinked?
- Widely used —> widely used,
- Please cite numpy, scipy, Jupyter notebook
- Please hyperlink GitHub and pypi for your specific thing
- Any citations for the applications?
- What about other pythonic radiative transfer codes?
- Nice clearly written statement of need!

Documentation
- Please also include the installation instructions here.
