### 2024-04-29 #6442 @dhjx1996
@arjunsavel Thank you very much for the review. I greatly appreciate the feedback and praise. Here are my updates with respect to your feedback:

> Thank you for your patience. I’ve finished my review of the PythonicDISORT package. It’s an excellent piece of work that has clear utility in both pedagogical and research contexts.
>
> My comments are below. Provided they’re addressed, I recommend this project for acceptance.
>
> Miscellaneous
>
> * This seems like quite a large GitHub repo. I believe 695 MB?
>
>   * Could some of the size be cut down by trimming notebook histories? And evaluate the notebooks on the fly?

The repository's large size is due to the .npz files for the pytest reference solutions. I have reduced the $\tau$ test points to exactly those used in the DISORT tests (`disotest.f90`) and the repository is now 140 MB. Let me know if this size reduction is sufficient.

> * The benefit of downloading the optional dependencies is not immediately apparent.

Made changes to README.md.

> * No community guidelines. Maybe make a contributing.md?

Added.

> * Really impressive set of notebooks!
> * Thoughts on including polarization?

It is not a priority especially since the feature is not in DISORT, but as a rule I am happy to add features to PythonicDISORT.

>
> Code edits
>
> * Could you please add function documentation in _one_Fourier_mode? For each argument?
> * Same with _loop_and_assemble_results. Even though these aren’t ever accessed directly, documenting these makes for easier contributor / developer experience.

I have improved the documentation in every module.

> * pydisort is a bit of a long function and could likely be modularized. Is it implemented this way to be more one-to-one with the original fortran? Either way, this isn’t something that gets in the way of acceptance — I’m mostly just curious.

This is caused by the length of the NT correction functions. I can modularize the NT corrections but I am currently disinclined to that change since it would be a major deviation from DISORT wherein NT corrections are under-the-hood and always performed if possible.

> * Some code could maybe be vectorized — e.g. in src/PythonicDISORT/_diagonalize.py. Again, not required for acceptance.

Yes, and I have and will continue to optimize the code. As of v0.7.0, however, I believe every expensive operation (e.g. diagonalization) is vectorized and no longer in Python for-loops.

> * Could you add a more descriptive docstring to _assemble_solution_functions?

Added.

>
> Paper edits
>
> * Can we get definitions of the terms in the equations?

Added.

> * Are omega and D functions of tau?

No. I added clarifications.

> * You mention computational reasons…are those temporary? Is speed optimization a future goal, e.g., with numba, or Jax, or python? Or even joblib for parallelization? If you end up using Jax, you could replace your autograd usage with jax.

Speed optimization has been and remains a goal, but it is impossible for the Python-based PythonicDISORT to match the FORTRAN-based DISORT in speed. I have attempted parallelization with joblib but the overhead actually results in a net slow down for all practical use cases. I am disinclined to use numba or Jax because they diminish the main advantage of PythonicDISORT which is its ease of use and installation. In fact, I have in CONTRIBUTING.md which states that "PythonicDISORT should remain system agnostic and installable through a simple `pip install` command. If your contributions have dependencies, they should be similarly accessible and easy to install."

> * Is “comprehensive documentation” meant to be hyperlinked?
> * Widely used —> widely used,
> * Please cite numpy, scipy, Jupyter notebook
> * Please hyperlink GitHub and pypi for your specific thing

These changes made been made.

> * Any citations for the applications?

Unfortunately I am not aware of any published applications at the moment (PythonicDISORT is quite new). At least two of the projects which are using PythonicDISORT are close to publication, however, and if they are published before the JOSS paper then I will add the citations.

> * What about other pythonic radiative transfer codes?

They are Python wrappers around DISORT. I have added two citations into the paper.

> * Nice clearly written statement of need!
>
> Documentation
>
> * Please also include the installation instructions here.

Added.
