### 2024-04-29 #6442 @dhjx1996
@phibeck @simonrp84 @arjunsavel @pscicluna The latest version of PythonicDISORT is now v0.7.0. This has been released on GitHub and on PyPI.
