### 2024-04-29 #6442 @dhjx1996
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1038/s41592-019-0686-2 is OK
> - 10.1038/s41586-020-2649-2 is OK
> - 10.1364/AO.27.002502 is OK
> - 10.1109/WHISPERS.2014.8077573 is OK
> - 10.1016/S0098-3004(97)00130-1 is OK
> - 10.1175/1520-0477(1998)079<2101:SARATS>2.0.CO;2 is OK
> - 10.1093/mnras/111.4.377 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Jupyter Notebooks – a publishing format for reprod...
> - No DOI given, and none found for title: Radiative Transfer
> - No DOI given, and none found for title: pyRT_DISORT: A pre-processing front-end to help ma...
> - No DOI given, and none found for title: pyDISORT
> - No DOI given, and none found for title: LLLab disort website
>
> INVALID DOIs
>
> - None
> ```

@phibeck Regarding the missing DOIs, these are all references to packages. The one exception is the book "Radiative Transfer" by Chandrasekhar (1960) for which I could not find a DOI.
