### 2024-05-01 #6442 @simonrp84
This is a very interesting library, thanks for creating it! Unfortunately I cannot test all of the code due to a missing module, for which I've created a github issue. This module is included in the github repo, but the docs do not describe how to install it.

I agree with @arjunsavel's comments on the manuscript and library, so please address all of those. In addition, I have some other points to raise - which are listed below.

**Paper comments:**

 - The software paper is well-written. However, I miss a brief introduction (one or two sentences) as to why an RT code is needed.

 - Throughout the paper there are various terms that are not introduced. For example, theta (line 9). Even though it's usually clear what these terms are from the context I think the terms should be explicitly introduced. From personal experience, different DISORT users have different interpretations of geometry terms, so being clear would be very helpful here.

 - Same for I_0 and mu_0, which are also undefined.

 - Line 33: ODE is not introduced.

 - Line 46: "projects which" should be "projects that"

 - General paper comment, I see some places where "we" is used (f.ex line 47: "We will continue...") this seems odd for a single-author paper. Either there should be more authors, or "we" is inappropriate. Can you please clarify this in the text and / or repository.

**Code/repo comments:**

 - It looks like the code in the `./disort4.0.99_f2py/` directory is based on code from elsewhere...does that affect the license?

 - The git repo is huge (~140Mb) compared to the size of the source code (~2.5Mb). I presume this is a remnant to the tidying that the other reviewer recommended. Can you do clean up the history to reduce this file size?

 - For the docs, could you also add a statement saying that the notebook should be used as an example? This is mentioned in the readme but not, as far as I can see, in the docs themselves.

 - The notebook in the docs folder is very comprehensive and nicely introduces the various concepts and uses of the library, great work!

 - I do miss a simple example, though. The notebook is somewhat overwhelming for a newbie. Would it be possible to add a notebook demonstrating some simple use of DISORT? Maybe simulating the signal for an airborne or spaceborne sensor, the effect of Rayleigh scattering, or something like that? I note in this context the reviewer checklist item: "Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)."

I've ticked the appropriate boxes in the reviewer checklist, and will await an author response before going further. Sorry again for my long delay in performing this review.
