### 2024-05-03 #6442 @dhjx1996
As a follow-up to my previous comment, I will revamp the current test notebooks into examples for users and remove `disort4.0.99_f2py` and all references to the `disort` module to avoid licensing issues. I may make some changes to how the tests are conducted but there will still be automated tests at the very least.
