### 2024-05-06 #6442 @phibeck
Hi @dhjx1996 I think this is a good idea to avoid licensing issues, although the details of such potential issues would probably depend on the license of DISORT. It seems sensible to me to remove the module and just provide the reference solutions as benchmarks. Of course that requires that the author(s) of PythonicDISORT keep this data up-to-date.

Thanks @simonrp84 for getting your review started! @pscicluna please let me know if you have any question about the review process or if you need more time as ideally we'd like to finish the review within 4-6 weeks.
