### 2024-06-04 91ba12e
Write-up for single-layer is complete. More exposition and code to add though
