+++ subroutines.py
@@ -1,0 +1,214 @@
+import numpy as np
+import scipy as sc
+from math import pi
+
+def transform_interval(arr, c, d, a=-1, b=1):
+    """Transforms a vector in interval [a, b] to a similar vector in [c, d]
+
+    :Input:
+     - *arr* (vector) - The vector to transform
+     - *c* (float) - The beginning of interval [c, d]
+     - *d* (float) - The end of interval [c, d]
+     :Optional:
+     - *a* (float) - The beginning of interval [a, b]
+     - *b* (float) - The end of interval [a, b]
+
+
+    :Output:
+     - (vector) - The transformed vector
+    """
+    return (((arr - a) * (d - c)) / (b - a)) + c
+
+
+def transform_weights(weights, c, d, a=-1, b=1):
+    """Transform quadrature weights in interval [a, b] to similar weights in [c, d]
+
+    :Input:
+     - *weights* (vector) - The weights to transform
+     - *c* (float) - The beginning of interval [c, d]
+     - *d* (float) - The end of interval [c, d]
+     :Optional:
+     - *a* (float) - The beginning of interval [a, b]
+     - *b* (float) - The end of interval [a, b]
+
+    :Output:
+     - *weights* (vector) - The transformed weights
+    """
+    return weights * (d - c) / (b - a)
+	
+	
+def calculate_nu(mu, phi, mu_p, phi_p):
+    """Calculates the scattering angle nu between incident angle (mu_p, phi_p) and scattering angle (mu, phi)
+    
+    :Input:
+     - *mu* (vector / float) - The weights to transform
+     - *phi* (vector / float) - The beginning of interval [c, d]
+     - *mu_p* (vector / float) - The end of interval [c, d]
+     - *phi_p* (vector / float) - The beginning of interval [c, d]
+     
+     
+    :Output:
+     - *nu* (tensor / float) - The transformed weights 
+    """
+    mu, phi, mu_p, phi_p = np.atleast_1d(mu, phi, mu_p, phi_p)
+    nu = mu_p[None, None, :, None] * mu[:, None, None, None] + np.sqrt(1 - mu_p**2)[
+        None, None, :, None
+    ] * np.sqrt(1 - mu**2)[:, None, None, None] * np.cos(
+        phi_p[None, None, None, :] - phi[None, :, None, None]
+    )
+    return np.squeeze(nu)
+
+
+def generate_Ds(m, Leg_coeffs, mu_arr_pos, w0, ells, degree_tile):
+    """Generates the D term in the system of ODEs for each Fourier mode, see Section 3.2
+    
+    :Input:
+     - *m* (integer) - The Fourier mode
+     - *Leg_coeffs* (vector) - Vector of weighted phase function Legendre coefficients
+     - *mu_arr_pos* (vector) - Positive mu (quadrature) values
+     - *w0* (float) - Single-scattering albedo
+     - *ells* (vector) - Vector of Legendre polynomial orders
+     - *degree_tile* (matrix) - Matrix of Associated Legendre polynomials orders and degrees
+     
+     
+    :Output:
+     - *D_pos* (matrix) - D with only elements that correspond to positive mu values
+     - *D_neg* (matrix) - D with only elements that correspond to negative mu values
+    """
+    Dm_term = Leg_coeffs[ells] * (
+        sc.special.factorial(ells - m) / sc.special.factorial(ells + m)
+    )
+
+    asso_leg_term_pos = sc.special.lpmv(m, degree_tile, mu_arr_pos)
+    asso_leg_term_neg = sc.special.lpmv(m, degree_tile, -mu_arr_pos)
+    D_temp = (Dm_term)[None, :] * asso_leg_term_pos.T
+
+    D_pos = (w0 / 2) * D_temp @ asso_leg_term_pos
[…]
