+++ README.md
@@ -1,1 +1,4 @@
 # Pythonic-DISORT
+
+Code to solve the radiative transfer equation for a single atmospheric layer using the discrete ordinates method in Python.
+Based on Stamnes' DISORT (see reference in Jupyter Notebook).
