### 2024-06-04 a08ef2b
PyDISORT packaged as module. Everything has been thoroughly verified and Stamnes' DISORT is fully working. Left the remaining TODOs.
