+++ PyDISORT/pydisort.py
@@ -1,19 +1,27 @@
-import numpy as np
+import PyDISORT
+try:
+    import autograd.numpy as np
+except ImportError:
+    import numpy as np
 import scipy as sc
+import warnings
+import PyDISORT
 from math import pi
 from numpy.polynomial import legendre
+from scipy import integrate
+from inspect import signature
 
-def PyDISORT(
-    b_pos, b_neg, only_flux, NQuad, NLeg, Leg_coeffs_full, tau0, w0, mu0, phi0, I0, f=0, p_for_NT_with_g=(None,),
+def pydisort(
+    b_pos, b_neg, only_flux, NQuad, NLoops, Leg_coeffs, tau0, w0, mu0, phi0, I0, f=0, p_for_NT=None,
 ):
     """Full radiative transfer solver which performs corrections
     
     :Input:
-     - *b_pos / neg* (matrix) - Boundary conditions for the upward / downward directions
+     - *b_pos / neg* (float matrix) - Boundary conditions for the upward / downward directions
      - *only_flux* (boolean) - Flag for whether to compute the intensity function
      - *NQuad* (integer) - Number of mu quadrature points
-     - *NLeg* (integer) - Number of Legendre coefficients to be used in the basic solver
-     - *Leg_coeffs_full* (vector) - Long vector of weighted phase function Legendre coefficients
+     - *NLoops* (integer) - Number of loops, also number of Fourier modes in the numerical solution
+     - *Leg_coeffs* (float vector) - Phase function Legendre coefficients
      - *tau0* (float) - Optical depth
      - *w0* (float) - Single-scattering albedo
      - *mu0* (float) - Polar angle of the direct beam
@@ -21,100 +29,102 @@ def PyDISORT(
      - *I0* (float) - Intensity of the direct beam
      :Optional:
      - *f* (float) - fractional scattering into the forward peak for use in delta-M scaling
-     - *p_for_NT_with_g* (tuple: (function, float)) - Tuple: (True phase function, asymmetry factor)
+     - *p_for_NT* (function) - True phase function with cosine of scattering angle as vector argument
      
      
     :Output:
-     - *mu_arr* (vector) - All mu values
+     - *mu_arr* (float vector) - All mu values
      - *flux_up* (function) - Flux function with argument tau for positive (upward) mu values
-     - *flux_down* (vector) - Flux function with argument tau for negative (downward) mu values
+     - *flux_down* (function) - Flux function with argument tau for negative (downward) mu values
      :Optional:
      - *u* (function) - Intensity function with arguments (tau, phi); the output is in the order (mu, tau, phi)
     """
-    Leg_coeffs = Leg_coeffs_full[:NLeg]
+    NLeg = len(Leg_coeffs)
     
     # INPUT CHECKS
     # -----------------------------------------------------------
-    # The fractional scattering into peak must be 0<= and <1
+    # Optical depth must be positive
+    assert tau0 > 0
+    # Single-scattering albedo must be between 0 and 1, excluding 1
+    assert w0 >= 0
+    assert w0 < 1
+    # There must be a positive number of Legendre coefficients each with magnitude <= 1
+    assert NLeg > 0
+    assert np.all(np.abs(Leg_coeffs) <= 1)
+    # Conditions on the number of quadrature angles (NQuad), Legendre coefficients (NLeg) and loops (NLoops)
+    assert NQuad >= 2
[…]
