+++ PyDISORT/subroutines.py
@@ -1,4 +1,7 @@
-import numpy as np
+try:
+    import autograd.numpy as np
+except ImportError:
+    import numpy as np
 import scipy as sc
 from math import pi
 
@@ -24,7 +27,7 @@ def transform_weights(weights, c, d, a=-1, b=1):
     """Transform quadrature weights in interval [a, b] to similar weights in [c, d]
 
     :Input:
-     - *weights* (vector) - The weights to transform
+     - *weights* (float vector) - The weights to transform
      - *c* (float) - The beginning of interval [c, d]
      - *d* (float) - The end of interval [c, d]
      :Optional:
@@ -41,10 +44,10 @@ def calculate_nu(mu, phi, mu_p, phi_p):
     """Calculates the scattering angle nu between incident angle (mu_p, phi_p) and scattering angle (mu, phi)
     
     :Input:
-     - *mu* (vector / float) - The weights to transform
-     - *phi* (vector / float) - The beginning of interval [c, d]
-     - *mu_p* (vector / float) - The end of interval [c, d]
-     - *phi_p* (vector / float) - The beginning of interval [c, d]
+     - *mu* (float vector) - Cosine of scattering polar angles
+     - *phi* (float vector) - Scattering azimuthal angles
+     - *mu_p* (float vector) - Cosine of incident polar angles
+     - *phi_p* (float vector) - Incident azimuthal angles
      
      
     :Output:
@@ -64,16 +67,16 @@ def generate_Ds(m, Leg_coeffs, mu_arr_pos, w0, ells, degree_tile):
     
     :Input:
      - *m* (integer) - The Fourier mode
-     - *Leg_coeffs* (vector) - Vector of weighted phase function Legendre coefficients
-     - *mu_arr_pos* (vector) - Positive mu (quadrature) values
+     - *Leg_coeffs* (float vector) - Vector of weighted phase function Legendre coefficients
+     - *mu_arr_pos* (float vector) - Positive mu (quadrature) values
      - *w0* (float) - Single-scattering albedo
-     - *ells* (vector) - Vector of Legendre polynomial orders
-     - *degree_tile* (matrix) - Matrix of Associated Legendre polynomials orders and degrees
+     - *ells* (float vector) - Vector of Legendre polynomial orders
+     - *degree_tile* (float matrix) - Matrix of Associated Legendre polynomials orders and degrees
      
      
     :Output:
-     - *D_pos* (matrix) - D with only elements that correspond to positive mu values
-     - *D_neg* (matrix) - D with only elements that correspond to negative mu values
+     - *D_pos* (float matrix) - D with only elements that correspond to positive mu values
+     - *D_neg* (float matrix) - D with only elements that correspond to negative mu values
     """
     Dm_term = Leg_coeffs[ells] * (
         sc.special.factorial(ells - m) / sc.special.factorial(ells + m)
@@ -89,13 +92,12 @@ def generate_Ds(m, Leg_coeffs, mu_arr_pos, w0, ells, degree_tile):
     return D_pos, D_neg
 	
 
-# Generate the X term in the system of ODEs
 def generate_Xs(m, Leg_coeffs, w0, mu0, I0, mu_arr_pos, ells, degree_tile):
     """Generates the X term in the system of ODEs for each Fourier mode, see Section 3.2
     
     :Input:
      - *m* (integer) - The Fourier mode
-     - *Leg_coeffs* (vector) - Vector of weighted phase function Legendre coefficients
+     - *Leg_coeffs* (vector) - WEIGHTED phase function Legendre coefficients
      - *w0* (float) - Single-scattering albedo
      - *mu0* (float) - Polar angle of the direct beam
      - *I0* (float) - Intensity of the direct beam
[…]

[…]
