+++ PyDISORT/__init__.py
@@ -1,3 +1,3 @@
 from PyDISORT import subroutines
-from PyDISORT.basic_solver import basic_solver
+from PyDISORT._basic_solver import _basic_solver
 from PyDISORT.pydisort import pydisort
