+++ PyDISORT/subroutines.py
@@ -4,54 +4,77 @@ except ImportError:
     import numpy as np
 import scipy as sc
 from math import pi
+from numpy.polynomial.legendre import leggauss
+
 
 def transform_interval(arr, c, d, a=-1, b=1):
-    """Transforms a vector in interval [a, b] to a similar vector in [c, d]
+    """Affine transformation of an array from interval [a, b] to [c, d].
 
-    :Input:
-     - *arr* (vector) - The vector to transform
-     - *c* (float) - The beginning of interval [c, d]
-     - *d* (float) - The end of interval [c, d]
-     :Optional:
-     - *a* (float) - The beginning of interval [a, b]
-     - *b* (float) - The end of interval [a, b]
+    Parameters
+    ----------
+    arr : array
+        The 1D array to transform.
+    c : float
+        The beginning of interval [c, d].
+    d : float
+        The end of interval [c, d].
+    a : float, optional
+        The beginning of interval [a, b].
+    b : float, optional
+        The end of interval [a, b].
 
+    Returns
+    -------
+    array
+        The transformed 1D array.
 
-    :Output:
-     - (vector) - The transformed vector
     """
     return (((arr - a) * (d - c)) / (b - a)) + c
 
-
 def transform_weights(weights, c, d, a=-1, b=1):
-    """Transform quadrature weights in interval [a, b] to similar weights in [c, d]
+    """Transforms quadrature weights from interval [a, b] to [c, d].
 
-    :Input:
-     - *weights* (float vector) - The weights to transform
-     - *c* (float) - The beginning of interval [c, d]
-     - *d* (float) - The end of interval [c, d]
-     :Optional:
-     - *a* (float) - The beginning of interval [a, b]
-     - *b* (float) - The end of interval [a, b]
+    Parameters
+    ----------
+    weights : array
+        The weights to transform.
+    c : float
+        The beginning of interval [c, d].
+    d : float
+        The end of interval [c, d].
+    a : float, optional
+        The beginning of interval [a, b].
+    b : float, optional
+        The end of interval [a, b].
 
-    :Output:
-     - *weights* (vector) - The transformed weights
+    Returns
+    -------
+    array
+        The transformed weights.
+
     """
     return weights * (d - c) / (b - a)
 	
 	
 def calculate_nu(mu, phi, mu_p, phi_p):
-    """Calculates the scattering angle nu between incident angle (mu_p, phi_p) and scattering angle (mu, phi)
-    
-    :Input:
-     - *mu* (float vector) - Cosine of scattering polar angles
-     - *phi* (float vector) - Scattering azimuthal angles
-     - *mu_p* (float vector) - Cosine of incident polar angles
-     - *phi_p* (float vector) - Incident azimuthal angles
-     
-     
-    :Output:
-     - *nu* (tensor / float) - The transformed weights 
+    """Calculates the scattering angle nu between incident angle (mu_p, phi_p) and scattering angle (mu, phi).
+
+    Parameters
+    ----------
+    mu : array
+        Cosine of outgoing polar angles.
+    phi : array
+        Outgoing azimuthal angles.
+    mu_p : array
+        Cosine of incident polar angles.
+    phi_p : array
+        Incident azimuthal angles.
+
+    Returns
+    -------
+    nu_arr : ndarray
+        Cosine of scattering angles.
+
     """
     mu, phi, mu_p, phi_p = np.atleast_1d(mu, phi, mu_p, phi_p)
[…]
