+++ docs/index.rst
@@ -1,0 +1,22 @@
+.. Pythonic-DISORT documentation master file, created by
+   sphinx-quickstart on Tue Mar 14 00:46:27 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Documentation for Pythonic-DISORT project
+=========================================
+
+Modules (Basic Documentation)
+=============================
+.. toctree::
+   :maxdepth: 4
+   
+   source/PyDISORT
+   
+Jupyter Notebook (Comprehensive Documentation)
+==============================================
+.. toctree::
+   :maxdepth: 2
+   
+   Pythonic-DISORT
+  
