+++ docs/source/PyDISORT.rst
@@ -1,0 +1,26 @@
+PyDISORT package
+================
+
+PyDISORT.pydisort module
+------------------------
+
+.. automodule:: PyDISORT.pydisort
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+PyDISORT.subroutines module
+---------------------------
+
+.. automodule:: PyDISORT.subroutines
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: PyDISORT
+   :members:
+   :undoc-members:
+   :show-inheritance:
