### 2024-06-04 aa14f19
Final few changes to docs before letting Prof. Pincus take a look :)
