+++ PyDISORT/pydisort.py
@@ -1,13 +1,14 @@
-import PyDISORT
-try:
-    import autograd.numpy as np
-except ImportError:
-    import numpy as np
+from PyDISORT import subroutines, _loop_and_assemble_results
 import scipy as sc
 from math import pi
 from numpy.polynomial.legendre import Legendre
 from scipy import integrate
 from inspect import signature
+try:
+    import autograd.numpy as np
+except ImportError:
+    import numpy as np
+
 
 def pydisort(
     tau_arr, omega_arr,
@@ -25,7 +26,10 @@ def pydisort(
     mathscr_vs=None,
     parfor_Fourier=False
 ):
-    """Full radiative transfer solver with corrections and input checks.
+    """Solves the 1D RTE for the fluxes, and optionally intensity,
+    of a multilayer atmosphere with the specified optical properties, boundary conditions
+    and sources. Optionally performs delta-M scaling and NT corrections.
+
 
     Parameters
     ----------
@@ -64,38 +68,45 @@ def pydisort(
         Must have arguments (tau_arr, int, 2darray, array, 2darray) and output (2darray)..
     parfor_Fourier : optional, bool
         Parallelize the for-loop over Fourier modes?
-        
+
     Returns
     -------
-    mu_arr : array
+    array
         All mu (cosine of polar angle) quadrature nodes.
-    flux_up : function
+    function
         Flux function with argument tau (type: array) for positive (upward) mu values.
         Returns the diffuse flux magnitudes (type: array).
-    flux_down : function
+    function
         Flux function with argument tau (type: array)  for negative (downward) mu values.
-        Returns a tuple of the diffuse and direct flux magnitudes respectively (type: (array, array)). 
-    u : function, optional
+        Returns a tuple of the diffuse and direct flux magnitudes respectively (type: (array, array)).
+    function, optional
         Intensity function with arguments (tau, phi) of types (array, array).
         The output is an ndarray with axes corresponding to (mu, tau, phi) variation.
 
     """
+    # Turn scalars to arrays
+    # ---------------------------------------------------------------------------------------------------------
     tau_arr = np.atleast_1d(tau_arr)
     omega_arr = np.atleast_1d(omega_arr)
     Leg_coeffs_all = np.atleast_2d(Leg_coeffs_all)
     f_arr = np.atleast_1d(f_arr)
-    
+
+    # Some setup
+    # ---------------------------------------------------------------------------------------------------------
     if NLeg == None:
         NLeg = NQuad
     if NLoops == None:
         NLoops = NQuad
+    NLeg_all = np.shape(Leg_coeffs_all)[1]
     Leg_coeffs = Leg_coeffs_all[:, :NLeg]
-    NLayers = len(tau_arr)
     scalar_b_pos, scalar_b_neg = False, False
     mathscr_vs_callable = callable(mathscr_vs)
-    
-    # INPUT CHECKS
-    # -----------------------------------------------------------
+    NLayers = len(tau_arr)
+    NBDRF = len(Leg_coeffs_BDRF)
+    weighted_Leg_coeffs_BDRF *= (2 * np.arange(NBDRF) + 1) * Leg_coeffs_BDRF
+
+    # Input checks
+    # ---------------------------------------------------------------------------------------------------------
     # Optical depth must be positive
     assert np.all(tau_arr > 0)
     # Single-scattering albedo must be between 0 and 1, excluding 1
@@ -105,15 +116,17 @@ def pydisort(
     # We must obviously supply more Legendre coefficients than we intend to use
[…]
