+++ PyDISORT/subroutines.py
@@ -86,96 +86,8 @@ def calculate_nu(mu, phi, mu_p, phi_p):
     return np.squeeze(nu)
 
 
-def generate_flux_functions(
-    tau_arr, 
-    I0, mu0,
-    GC_pos, GC_neg, 
-    eigenvals, N,
-    B_pos, B_neg, 
-    mu_arr_pos, weights_mu, 
-    scale_tau,
-):
-    """Generates the flux functions with respect to the radiative transfer equation.
-
-    Parameters
-    ----------
-    tau_arr : array
-        Optical depth of the lower boundary of each atmospheric layer.
-    I0 : float
-        Intensity of the incident beam.
-    mu0 : float
-        Polar angle of the incident beam.
-    GC_pos : 2darray
-        Product of eigenvectors and coefficients that correspond to positive mu values.
-    GC_neg : 2darray
-        Product of eigenvectors and coefficients that correspond to negative mu values.
-    eigenvals : array
-        Eigenvalues arranged negative then positive, from largest to smallest magnitude.
-    N : int
-        Half the number of quadrature points.
-    B_pos : array
-        Coefficients of the incident beam inhomogeneity that correspond to positive mu values.
-    B_neg : array
-        Coefficients of the incident beam inhomogeneity that correspond to negative mu values.
-    mu_arr_pos : array
-        Positive mu quadrature nodes.
-    weights_mu : array
-        mu quadrature weights for positive mu nodes.
-    scale_tau : array
-        Delta-M scale factors for tau.
-
-    Returns
-    -------
-    function
-        Flux function with argument tau (type: array) for positive (upward) mu values.
-        Returns diffuse flux magnitudes (type: array).
-    function
-        Flux function with argument tau (type: array)  for negative (downward) mu values.
-        Returns a tuple of diffuse and direct flux magnitudes respectively (type: (array, array)).
-
-    """
-
-    def flux_up(tau):
-        tau = scale_tau * np.atleast_1d(tau)  # Delta-M scaling
-        exponent = np.vstack(
-            (
-                eigenvals[:N, None] * tau[None, :],
-                eigenvals[N:, None] * (tau - taul)[None, :],
-            )
-        )
-        u0_pos = GC_pos @ np.exp(exponent) + B_pos[:, None] * np.exp(
-            -tau[None, :] / mu0
-        )
-        return np.squeeze(2 * pi * (mu_arr_pos * weights_mu) @ u0_pos)[()]
-
-    def flux_down(tau):
-        direct_beam = I0 * mu0 * np.exp(-tau / mu0)
-
-        tau = scale_tau * np.atleast_1d(tau)  # Delta-M scaling
-        direct_beam_scaled = I0 * mu0 * np.exp(-tau / mu0)
-        exponent = np.vstack(
-            (
-                eigenvals[:N, None] * tau[None, :],
-                eigenvals[N:, None] * (tau - taul)[None, :],
-            )
-        )
-        u0_neg = GC_neg @ np.exp(exponent) + B_neg[:, None] * np.exp(
-            -tau[None, :] / mu0
-        )
-        return (
-            np.squeeze(
-                2 * pi * (mu_arr_pos * weights_mu) @ u0_neg
-                + direct_beam_scaled
-                - direct_beam
-            )[()],
-            direct_beam,
-        )
-
-    return flux_up, flux_down
-
-
 def Gauss_Legendre_quad(N, c=0, d=1):
[…]
