+++ PyDISORT/_one_Fourier_mode.py
@@ -1,182 +1,353 @@
-from PyDISORT import subroutines
+from PyDISORT.subroutines import _mathscr_v
 import numpy as np
 import scipy as sc
 from math import pi
 
+
 def _one_Fourier_mode(
     m,
-    tau_arr, omega_arr,
-    N, NQuad, NLeg,
-    weighted_Leg_coeffs,
-    mu0, I0, phi0,
-    b_pos, b_neg,
-    NLayers, NBDRF,
-    weighted_Leg_coeffs_BDRF,
-    mathscr_vs,
-    mu_arr_pos, weights_mu,
+    scaled_omega_arr,
+    tau_arr,
+    scaled_tau_arr_with_0,
+    mu_arr_pos, mu_arr,
     M_inv, W,
-    scale_tau
-):  
+    N, NQuad, NLeg,
+    NLayers, NBDRF,
+    weighted_scaled_Leg_coeffs,
+    weighted_Leg_coeffs_BDRF,
+    mu0, I0,
+    b_pos, b_neg,
+    scalar_b_pos, scalar_b_neg,
+    s_poly_coeffs,
+    Nscoeffs,
+    use_sparse_NLayers
+):
     """This function is wrapped and looped by the `_loop_and_assemble_results` function.
     It has many seemingly redundant arguments to maximize precomputation in the `pydisort` function.
     See the Jupyter Notebook, especially section 3, for documentation, explanation and derivation.
-    The labels in this file reference labels in the Jupyter Notebook, especially section 3.
-    
+    The labels in this file reference labels in the Jupyter Notebook, especially sections 3 and 5.
+
     """
     # Setup
     # --------------------------------------------------------------------------------------------------------------------------
     ells = np.arange(m, NLeg)
     degree_tile = np.tile(ells, (N, 1)).T
-    fac = np.prod(ells[:, None] + np.arange(-m + 1, m + 1)[None, :], axis=-1)
+    fac = np.prod(
+        ells[:, None] + np.arange(-m + 1, m + 1)[None, :], axis=-1, dtype=np.float64
+    )
     signs = np.empty(NLeg - m)
     signs[::2] = 1
     signs[1::2] = -1
-    if m == 0:
-        prefactor = omegal * I0 / (4 * pi)
-    else:
-        prefactor = omegal * I0 / (2 * pi)
-        
+
     asso_leg_term_pos = sc.special.lpmv(m, degree_tile, mu_arr_pos)
     asso_leg_term_neg = asso_leg_term_pos * signs[:, None]
     asso_leg_term_mu0 = sc.special.lpmv(m, ells, -mu0)
-
+    # --------------------------------------------------------------------------------------------------------------------------
+    
     # Generate mathscr_D and mathscr_X (BDRF terms)
     # --------------------------------------------------------------------------------------------------------------------------
     # If h_\ell = 0 for all \ell \geq m, then there is no BDRF contribution
     if m < NBDRF:
-        if m == 0:
-            prefactor_mathscr = mu0 * I0 / pi
-        else:
-            prefactor_mathscr = mu0 * I0 * 2 / pi
-
         weighted_asso_Leg_coeffs_BDRF = (
             weighted_Leg_coeffs_BDRF[ells[: (NBDRF - m)]] / fac[: (NBDRF - m)]
         )
 
         mathscr_D_temp = (
-            weighted_asso_Leg_coeffs_BDRF[None, :] * asso_leg_term_pos[:NBDRF, :].T * mu_arr_pos[:, None]
+            weighted_asso_Leg_coeffs_BDRF[None, :] * asso_leg_term_pos.T[:, :NBDRF]
[…]
