+++ PyDISORT/README.md
@@ -1,9 +1,0 @@
-Solves the (1D) Radiative Transfer Equation for a multilayer atmospheric layer using the Discrete Ordinates Method. 
-Coded entirely in Python. 
-Based on Stamnes' DISORT (see reference in Jupyter Notebook) and has its main features: e.g. delta-M scaling, Nakajima-Tanaka (NT) corrections, only flux option, blackbody radiation source etc.
-
-Documentation: https://pythonic-disort.readthedocs.io/en/latest/.
-Find the accompanying Jupyter Notebook (Pythonic-DISORT.ipynb) at https://github.com/Maths2209/Pythonic-DISORT.
-The Jupyter Notebook provides comprehensive documentation, suggested inputs, explanations, 
-mathematical derivations and verification tests (separate from the tests implemented using pytest).
-We HIGHLY recommend reading sections 1 and 2, though the optional parts can be safely skipped.
