+++ docs/index.rst
@@ -15,8 +15,10 @@ Modules (Basic Documentation)
    
 Jupyter Notebook (Comprehensive Documentation)
 ==============================================
+We highly recommend reading the non-optional parts of sections 1 and 2 before use.
+
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 4
    
    Pythonic-DISORT
   
