+++ docs/index.rst
@@ -3,7 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Documentation for Pythonic-DISORT project
+Documentation for Pythonic DISORT project
 =========================================
 
 Modules (Basic Documentation)
