+++ pyproject.toml
@@ -4,15 +4,15 @@ build-backend = "setuptools.build_meta"
 
 [project]
 dependencies = [
-  'numpy>=1.12',
-  'scipy>=1.2.1',
+  "numpy>=1.12",
+  "scipy>=1.2.1",
 ]
 name = "PyDISORT"
-version = "0.0.0"
+version = "0.0.1"
 authors = [
   { name="Dion Ho", email="dh3065@columbia.edu" },
 ]
-description = "Solves the (1D) Radiative Transfer Equation for a multilayer atmospheric layer using the Discrete Ordinates Method."
+description = "Discrete Ordinates Solver for the (1D) Radiative Transfer Equation in a single or multi-layer atmosphere."
 readme = "README.md"
 requires-python = ">=3.7"
 classifiers = [
@@ -22,11 +22,13 @@ classifiers = [
 ]
 
 [project.optional-dependencies]
+pytest = ["pytest >= 6.2.5"]
 notebook_dependencies = [
-	'autograd>=1.5',
-	'jupyter>1.0.0',
-	'notebook>6.5.2',
+	"autograd>=1.5",
+	"jupyter>1.0.0",
+	"notebook>6.5.2",
+	"ipympl>=0.8.8",
 ]
 
 [project.urls]
-"Homepage" = "https://github.com/Maths2209/Pythonic-DISORT"
+"Homepage" = "https://github.com/LDEO-CREW/Pythonic-DISORT"
