### 2024-06-04 8ec266f
Continuous (automated) testing and distribution completed. Renamed package to "PythonicDISORT". The function name remains "PyDISORT". Ready to release v0.0.1. (PyTests omitted in history re-write)
