+++ src/PythonicDISORT/__init__.py
@@ -1,0 +1,2 @@
+from PythonicDISORT import subroutines
+from PythonicDISORT.pydisort import pydisort
