+++ src/PyDISORT/__init__.py
@@ -1,2 +1,0 @@
-from PyDISORT import subroutines
-from PyDISORT.pydisort import pydisort
