+++ docs/source/PyDISORT.rst
@@ -1,18 +1,0 @@
-PyDISORT package
-================
-
-PyDISORT.pydisort module
-------------------------
-
-.. automodule:: PyDISORT.pydisort
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-PyDISORT.subroutines module
----------------------------
-
-.. automodule:: PyDISORT.subroutines
-   :members:
-   :undoc-members:
-   :show-inheritance:
