+++ README.md
@@ -1,13 +1,13 @@
 # Introduction
-The PyDISORT package is a Discrete Ordinates Solver for the (1D) Radiative Transfer Equation in a single or multi-layer atmosphere.
+The PythonicDISORT package is a Discrete Ordinates Solver for the (1D) Radiative Transfer Equation in a single or multi-layer atmosphere.
 It is coded entirely in Python 3 and in as "Pythonic" a manner as possible: we vectorize as well as use list comprehension and `map` as much as possible. 
 
-It is based on Stamnes' FORTRAN DISORT (see references in the Jupyter Notebook) and has its main features: 
+It is based on Stamnes' FORTRAN DISORT (see References, in particular [2, 3, 8]) and has its main features: 
 delta-M scaling, Nakajima-Tanaka (NT) corrections, only flux option, isotropic internal sources (thermal source),
-Bi-Direction Reflectance Function (BDRF) and more.
+Bi-Directional Reflectance Function (BDRF) and more.
 
 This repository also includes our F2PY-wrapped Stamnes DISORT (version 4.0.99) in the `disort4.0.99_f2py` directory.
-The original was downloaded from http://www.rtatmocn.com/disort/.
+The original was downloaded from http://www.rtatmocn.com/disort/. The wrapper was inspired by https://github.com/kconnour/pyRT_DISORT.
 
 # Documentation
 https://pythonic-disort.readthedocs.io/en/latest/
@@ -24,13 +24,14 @@ With PyTest installed, execute the console command `pytest` in the `pydisotest` 
 
 # Installation
 
-* From PyPI: (TODO: need to first publish on PyPI)
+* From PyPI: `pip install PythonicDISORT`
 * From Conda-forge: (TODO: need to first publish on Conda-forge)
-* By cloning repository: `pip install .` in the `Pythonic-DISORT` directory; `pip install -r optional_dependencies.txt` to install all optional dependencies
+* By cloning repository: `pip install .` in the `Pythonic-DISORT` directory; `pip install -r all_optional_dependencies.txt` to install all optional dependencies
 
-## Requirements to run PyDISORT
-* `numpy >= 1.12`
-* `scipy >= 1.2.1`
+## Requirements to run PythonicDISORT
+* Python 3.8+
+* `numpy >= 1.17.3`
+* `scipy >= 1.8.0`
 * (OPTIONAL) `pytest >= 6.2.5`
 
 ## Additional requirements to run the Jupyter Notebook
@@ -43,7 +44,31 @@ Our F2PY-wrapped Stamnes' DISORT (in the `disort4.0.99_f2py` directory) must als
 
 ## Compatibility
 
-The PyDISORT package should be system agnostic given its minimal dependencies and pure Python code.
+The PythonicDISORT package should be system agnostic given its minimal dependencies and pure Python code.
 We are not sure if the Jupyter Notebook or our F2PY-wrapped Stamnes' DISORT will run without issues on non-Windows systems.
 Everything was built and tested on Windows 11 and not yet tested on other systems.
 
+# References
+1) S. Chandrasekhar. 1960. *Radiative Transfer.*
+
+2) Knut Stamnes and S-Chee Tsay and Warren Wiscombe and Kolf Jayaweera. 1988. *Numerically stable algorithm for discrete-ordinate-method radiative transfer in multiple scattering and emitting layered media.* http://opg.optica.org/ao/abstract.cfm?URI=ao-27-12-2502.
+
+3) Stamnes, S.. 1999. *LLLab disort website.* http://www.rtatmocn.com/disort/.
+
[…]
