+++ PythonicDISORT_optional_dependencies.txt
@@ -1,0 +1,3 @@
+# Optional dependencies for the PythonicDISORT package
+# Install through "pip install -r PythonicDISORT_optional_dependencies.txt"
+pytest
