+++ optional_dependencies.txt
@@ -1,6 +1,0 @@
-# All optional dependencies for the Pythonic-DISORT project
-pytest
-autograd
-jupyter
-notebook
-ipympl
