+++ all_optional_dependencies.txt
@@ -1,0 +1,7 @@
+# All optional dependencies for the Pythonic DISORT project
+# Install through "pip install -r all_optional_dependencies.txt"
+pytest
+autograd
+jupyter
+notebook
+ipympl
