+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. Pythonic-DISORT documentation master file, created by
+.. PythonicDISORT documentation master file, created by
    sphinx-quickstart on Tue Mar 14 00:46:27 2023.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Documentation for Pythonic DISORT project
+Documentation for Pythonic DISORT package
 =========================================
 
 Modules (Basic Documentation)
@@ -11,7 +11,7 @@ Modules (Basic Documentation)
 .. toctree::
    :maxdepth: 4
    
-   source/PyDISORT
+   source/PythonicDISORT
    
 Jupyter Notebook (Comprehensive Documentation)
 ==============================================
@@ -20,5 +20,5 @@ We highly recommend reading the non-optional parts of sections 1 and 2 before us
 .. toctree::
    :maxdepth: 4
    
-   Pythonic-DISORT
+   Pythonic DISORT
   
