### 2024-06-04 9394039
A bunch of docs and docstring changes. This leads to our 0.0.2 release and will be our final commit for awhile. (PyTests omitted in history re-write)
