+++ docs/conf.py
@@ -6,10 +6,10 @@
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'PythonicDISORT'
+project = 'Pythonic DISORT'
 copyright = '2023, HO Jia Xu Dion'
-author = 'Dion Ho JX'
-release = '0.0.0'
+author = 'Dion HO Jia Xu'
+release = '0.0.2'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
