+++ pyproject.toml
@@ -8,9 +8,9 @@ dependencies = [
   "scipy>=1.8.0",
 ]
 name = "PythonicDISORT"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
-  { name="Dion Ho", email="dh3065@columbia.edu" },
+  { name="Dion HO Jia Xu", email="dh3065@columbia.edu" },
 ]
 description = "Discrete Ordinates Solver for the (1D) Radiative Transfer Equation in a single or multi-layer atmosphere."
 readme = "README.md"
