### 2024-06-04 6356f63
Changed tests, ready for v0.1.0 release
Prevented overflow in the associated Legendre coefficients (changes made to src/PythonicDISORT/_one_Fourier_mode.py).
Changing that part of the code also increased speed by about 2ms (about 10%) and increased accuracy (still being tested).

Jupyter Notebook: changed 3D plots to contour plots

Removed verification test for reflectance and transmittance in the notebook
