+++ src/PythonicDISORT/_loop_and_assemble_results.py
@@ -1,4 +1,3 @@
-from PythonicDISORT import subroutines
 from PythonicDISORT.subroutines import _mathscr_v
 from PythonicDISORT._one_Fourier_mode import _one_Fourier_mode 
 from math import pi
