### 2024-06-04 4e59a09
Added parallel solving of Fourier modes using the joblib package. Ready to release v0.2.0 (PyTests omitted in history re-write)
