+++ src/PythonicDISORT/_loop_and_assemble_results.py
@@ -23,7 +23,8 @@ def _loop_and_assemble_results(
     Nscoeffs,
     scale_tau,
     only_flux,
-    use_sparse_NLayers
+    use_sparse_NLayers,
+    n_jobs
 ):  
     """This function is wrapped by the `pydisort` function.
     It should be called through `pydisort` and never directly.
@@ -77,10 +78,35 @@ def _loop_and_assemble_results(
         if I0 > 0:
             B_collect = np.empty((NLoops, NLayers, NQuad))
             B_collect[0, :, :] = B_collect_0
-        # TODO: Look into the "xarray.apply_ufunc" method or "dask delayed / bag" to parallelize this
-        # May need to code an xarray wrapper. Dask works with Python code.
-        for m in range(1, NLoops): 
-            outputs = _one_Fourier_mode(
+                
+        if n_jobs == 1:
+            for m in range(1, NLoops): 
+                outputs = _one_Fourier_mode(
+                    m,
+                    scaled_omega_arr,
+                    tau_arr,
+                    scaled_tau_arr_with_0,
+                    mu_arr_pos, mu_arr,
+                    M_inv, W,
+                    N, NQuad, NLeg,
+                    NLayers, NBDRF,
+                    weighted_scaled_Leg_coeffs,
+                    weighted_Leg_coeffs_BDRF,
+                    mu0, I0,
+                    b_pos, b_neg,
+                    scalar_b_pos, scalar_b_neg,
+                    s_poly_coeffs,
+                    Nscoeffs,
+                    use_sparse_NLayers
+                )
+                if I0 > 0:
+                    GC_collect[m, :, :, :], K_collect[m, :, :], B_collect[m, :, :] = outputs
+                else:
+                    GC_collect[m, :, :, :], K_collect[m, :, :] = outputs
+        else:
+            from joblib import Parallel, delayed
+
+            _one_Fourier_mode_m = lambda m: _one_Fourier_mode(
                 m,
                 scaled_omega_arr,
                 tau_arr,
@@ -98,10 +124,16 @@ def _loop_and_assemble_results(
                 Nscoeffs,
                 use_sparse_NLayers
             )
-            if I0 > 0:
-                GC_collect[m, :, :, :], K_collect[m, :, :], B_collect[m, :, :] = outputs
-            else:
-                GC_collect[m, :, :, :], K_collect[m, :, :] = outputs
+            outputs = Parallel(n_jobs=n_jobs)(
+                delayed(_one_Fourier_mode_m)(m) for m in range(1, NLoops)
+            )
+            
+            for m in range(1, NLoops):
+                outputs_m = outputs[m - 1]
+                GC_collect[m, :, :, :] = outputs_m[0]
+                K_collect[m, :, :] = outputs_m[1]
+                if I0 > 0:
+                    B_collect[m, :, :] = outputs_m[2]
                 
         # --------------------------------------------------------------------------------------------------------------------------
         
