### 2024-06-04 bf9fd51
Added some parallelization details to the Jupyter Notebook
Changes to pytests. Also added some information about the omega = 1 special case.

Change: check for whether single-scattering albedo omega is too close to 1 only for m == 0 else the checks will add too much runtime

Ready to release v0.2.1 (PyTests omitted in history re-write)
