### 2024-06-04 740724e
Made changes to help deal with large I0 inputs. Ready to release version 0.2.2. (PyTests omitted in history re-write)
