### 2024-06-04 bf628d3
Bugfixes. Ready to release v0.2.3 which is just v0.2.2 but with bugs fixed.
