### 2024-06-04 3c98d6e
We no longer set values close to zero to exactly zero as that caused more problems than it solved. We replaced all "np.allclose" and "np.isclose" from the algorithm to speed it up. Ready to release version 0.2.4. (PyTests omitted in history re-write)
