### 2024-06-04 fef99e5
Implemented a basic function to generate the diffuse actinic flux functions from u0. Ready to release v0.3.0. (PyTests omitted in history re-write)
