### 2024-06-04 2cf9214
Upgraded BDRF feature. Now to speed up the loop over each atmospheric layer.
Sped up loops over layers and changed timing tests to better demonstrate the benefits of parallelization and the use of sparse matrices.
Also increased the default `use_sparse_NLayers` from 6 to 13. Ready to release v0.4.0 (PyTests omitted in history re-write)
