### 2024-06-04 d6f551b
Corrected documentation to clarify that mu0 is the COSINE of the polar angle, and not just the polar angle
Upgraded subroutines.generate_diff_act_flux_funcs to automatically perform the reclassification of delta-scaled actinic flux

Messed up release of v0.4.1 so it will be skipped and v0.4.2 will take its place (PyTests omitted in history re-write)
