### 2024-06-04 53bbfde
Added draft of JOSS paper and made minor changes to README
Final submit before publication to JOSS

Named draft-pdf.yml action

Computation of eigenpairs is now completely vectorized (no for loops)

Made changes to README and other docs. Ready to release version 0.5.0.
