+++ docs/JOSS_Paper/paper.bib
@@ -1,0 +1,78 @@
+@article{STWJ1988,
+author = {Knut Stamnes and S-Chee Tsay and Warren Wiscombe and Kolf Jayaweera},
+journal = {Appl. Opt.},
+keywords = {Electromagnetic radiation; Multiple scattering; Optical depth; Radiative transfer; Reflection; Thermal emission},
+number = {12},
+pages = {2502--2509},
+publisher = {Optica Publishing Group},
+title = {Numerically stable algorithm for discrete-ordinate-method radiative transfer in multiple scattering and emitting layered media},
+volume = {27},
+month = {Jun},
+year = {1988},
+url = {http://opg.optica.org/ao/abstract.cfm?URI=ao-27-12-2502},
+doi = {10.1364/AO.27.002502},
+abstract = {We summarize an advanced, thoroughly documented, and quite general purpose discrete ordinate algorithm for time-independent transfer calculations in vertically inhomogeneous, nonisothermal, plane-parallel media. Atmospheric applications ranging from the UV to the radar region of the electromagnetic spectrum are possible. The physical processes included are thermal emission, scattering, absorption, and bidirectional reflection and emissionat the lower boundary. The medium may be forced at the top boundary by parallel or diffuse radiation and by internal and boundary thermal sources as well. We provide a brief account of the theoretical basis as well as a discussion of the numerical implementation of the theory. The recent advances made by ourselves and our collaborators---advances in both formulation and numerical solution---are all incorporated in the algorithm. Prominent among these advances are the complete conquest of two ill-conditioning problems which afflicted all previous discrete ordinate implementations: (1) the computation of eigenvalues and eigenvectors and (2) the inversion of the matrix determining the constants of integration. Copies of the fortran program on microcomputer diskettes are available for interested users.},
+}
+
+@misc{Sta1999, 
+	title={LLLab disort website}, 
+	url={http://www.rtatmocn.com/disort/}, 
+	journal={Light and Life Lab (LLLab)}, 
+	author={Stamnes, S.}, 
+	year={1999}
+}
+
+@INPROCEEDINGS{Ber2014,
+  author={Berk, A. and Conforti, P. and Kennett, R. and Perkins, T. and Hawes, F. and van den Bosch, J.},
+  booktitle={2014 6th Workshop on Hyperspectral Image and Signal Processing: Evolution in Remote Sensing (WHISPERS)}, 
+  title={MODTRAN® 6: A major upgrade of the MODTRAN® radiative transfer code}, 
+  year={2014},
+  volume={},
+  number={},
+  pages={1-4},
+  doi={10.1109/WHISPERS.2014.8077573}}
+
+@article{Key1998,
+title = {Tools for atmospheric radiative transfer: Streamer and FluxNet},
+journal = {Computers & Geosciences},
+volume = {24},
+number = {5},
+pages = {443-451},
+year = {1998},
+issn = {0098-3004},
+doi = {10.1016/S0098-3004(97)00130-1},
+url = {https://www.sciencedirect.com/science/article/pii/S0098300497001301},
+author = {Jeffrey R Key and Axel J Schweiger},
+keywords = {Radiative transfer models, Atmospheric processes, Radiation budget, Satellites},
[…]
