+++ docs/JOSS_Paper/paper.md
@@ -1,0 +1,157 @@
+---
+title: '`PythonicDISORT`: A Python reimplementation of the Discrete Ordinate Radiative Transfer package `DISORT`'
+tags:
+  - Python
+  - Radiative Transfer
+  - Discrete Ordinates Method
+  - Atmospheric Science
+  - Climate Models
+  - DISORT
+authors:
+  - name: Dion J. X. Ho
+    orcid: 0009-0000-5829-5081
+    affiliation: "1"
+affiliations:
+ - name: Columbia University, Department of Applied Physics and Applied Mathematics, USA
+   index: 1
+date: 11 February 2024
+bibliography: paper.bib
+---
+
+<!---
+title: 'Gala: A Python package for galactic dynamics'
+tags:
+  - Python
+  - astronomy
+  - dynamics
+  - galactic dynamics
+  - milky way
+authors:
+  - name: Adrian M. Price-Whelan
+    orcid: 0000-0000-0000-0000
+    equal-contrib: true
+    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+  - name: Author Without ORCID
+    equal-contrib: true # (This is how you can denote equal contributions between multiple authors)
+    affiliation: 2
+  - name: Author with no affiliation
+    corresponding: true # (This is how to denote the corresponding author)
+    affiliation: 3
+  - given-names: Ludwig
+    dropping-particle: van
+    surname: Beethoven
+    affiliation: 3
+affiliations:
+ - name: Lyman Spitzer, Jr. Fellow, Princeton University, USA
+   index: 1
+ - name: Institution Name, Country
+   index: 2
+ - name: Independent Researcher, Country
+   index: 3
+date: 13 August 2017
+bibliography: paper.bib
+
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-journal: Astrophysical Journal <- The name of the AAS journal.
+--->
+
+# Summary
+The Radiative Transfer Equation (RTE) models the processes of absorption, scattering and emission 
+as electromagnetic radiation propagates through a medium. We address the 1D RTE (\ref{RTE}) 
+in a plane-parallel atmosphere and consider three sources: 
+blackbody emission from the atmosphere $s(\tau)$, scattering from sunlight
+$\frac{\omega I_0}{4 \pi} p\left(\mu, \phi ;-\mu_{0}, \phi_{0}\right) \exp\left(-\mu_{0}^{-1} \tau\right)$,
+and incoming radiation from other atmospheric layers or the Earth's surface modeled by
+Dirichlet boundary conditions.
+
+\begin{align}
+\begin{split}
+\mu \frac{\partial u(\tau, \mu, \phi)}{\partial \tau} = u(\tau, \mu, \phi) &-\frac{\omega}{4 \pi} \int_{-1}^{1} \int_{0}^{2 \pi} p\left(\mu, \phi ; \mu', \phi'\right) u\left(\tau, \mu', \phi'\right) \mathrm{d} \phi' \mathrm{d} \mu' \\
+&-\frac{\omega I_0}{4 \pi} p\left(\mu, \phi ;-\mu_{0}, \phi_{0}\right) \exp\left(-\mu_{0}^{-1} \tau\right) - s(\tau)
+\end{split} \label{RTE}
+\end{align}
+
+The RTE is important in many fields of science and engineering.
+The gold standard for numerically solving the 1D RTE is the Discrete Ordinate Radiative Transfer 
+package `DISORT` which was coded in FORTRAN 77 and first released in 1988 [@STWJ1988].
[…]
