+++ all_optional_dependencies.txt
@@ -1,7 +1,6 @@
 # All optional dependencies for the Pythonic DISORT project
 # Install through "pip install -r all_optional_dependencies.txt"
 
-joblib
 pytest
 autograd
 jupyter
