### 2024-06-04 9c69346
Update continuous_distribution.yml
Modified sparse framework. Commit before completing changes

Made substantial optimizations. Ready for release of v0.6.0. (PyTests omitted in history re-write)
