### 2024-06-04 1f56080
Replaced `flatten` with `ravel`
Release v0.6.1 ready. We fixed a bug and reverted a change from the previous release (`np.add.at` is not fast enough). (PyTests omitted in history re-write)
