### 2024-06-04 43ba8e3
Renamed "NLoops" to "NFourier". Ready to release v0.6.2.
