### 2024-06-04 db77661
Completed code edits for v0.7.0 (and to address JOSS reviewer feedback)
Added CONTRIBUTING.md and made changes to documentation (PyTests omitted in history re-write)
