+++ src/PythonicDISORT/pydisort.py
@@ -1,5 +1,5 @@
 from PythonicDISORT import subroutines
-from PythonicDISORT._assemble_solution_functions import _assemble_solution_functions
+from PythonicDISORT._assemble_intensity_and_fluxes import _assemble_intensity_and_fluxes
 
 import numpy as np
 import scipy as sc
@@ -22,8 +22,8 @@ def pydisort(
     NT_cor=False,
     BDRF_Fourier_modes=[],
     s_poly_coeffs=np.array([[]]),
-    use_sparse_NLayers=13,
-    _is_compatible_with_autograd = False,
+    NLayers_to_use_sparse=13,
+    _autograd_compatible = False,
 ):
     """Solves the 1D RTE for the fluxes, and optionally intensity,
     of a multi-layer atmosphere with the specified optical properties, boundary conditions
@@ -37,14 +37,15 @@ def pydisort(
 
     Parameters
     ----------
-    tau_arr : array
+    tau_arr : array or float
         Optical depth of the lower boundary of each atmospheric layer.
-    omega_arr : array
+    omega_arr : array or float
         Single-scattering albedo of each atmospheric layer.
     NQuad : int
         Number of `mu` quadrature nodes.
-    Leg_coeffs_all : ndarray
-        All available unweighted phase function Legendre coefficients. 
+    Leg_coeffs_all : 2darray
+        All available unweighted phase function Legendre coefficients.
+        Each row of coefficients pertain to an atmospheric layer.
         Each coefficient should be between 0 and 1 inclusive.
     mu0 : float
         Cosine of polar angle of the incident beam.
@@ -53,9 +54,10 @@ def pydisort(
     phi0 : float
         Azimuthal angle of the incident beam.
     NLeg : optional, int
-        Number of phase function Legendre coefficients.
+        Number of phase function Legendre coefficients
+        to use in the pre-correction solver.
     NFourier : optional, int
-        Number of Fourier modes to use in the intensity function.
+        Number of Fourier modes to use to construct the intensity function.
     b_pos : optional, 2darray or float
         Dirichlet boundary condition for the upward direction.
     b_neg : optional, 2darray or float
@@ -72,33 +74,80 @@ def pydisort(
     s_poly_coeffs : optional, array
         Polynomial coefficients of isotropic internal sources.
         Arrange coefficients from lowest order term to highest.
-    use_sparse_NLayers : optional, int
+    NLayers_to_use_sparse : optional, int
         At or above how many atmospheric layers should SciPy's sparse matrix framework be used?
 
     Returns
     -------
-    array
+    mu_arr : array
         All `mu` (cosine of polar angle) quadrature nodes.
-    function
+    Fp(tau) : function, function
         (Energetic) Flux function with argument `tau` (type: array) for positive (upward) `mu` values.
         Returns the diffuse flux magnitudes (type: array).
-    function
+    Fm(tau) : function
         (Energetic) Flux function with argument `tau` (type: array) for negative (downward) `mu` values.
         Returns a tuple of the diffuse and direct flux magnitudes respectively (type: (array, array)).
-    function
+    u0(tau) : function
         Zeroth Fourier mode of the intensity with argument `tau` (type: array).
         Returns an ndarray with axes corresponding to variation with `mu` and `tau` respectively.
         This function is useful for calculating actinic fluxes and other quantities of interest,
         but reclassification of delta-scaled flux and other corrections must be done manually
         (for actinic flux `generate_diff_act_flux_funcs` will automatically perform the reclassification).
-    function, optional
+    u(tau, phi) : function, optional
         Intensity function with arguments `(tau, phi, return_Fourier_error=False)` of types `(array, array, bool)`.
         Returns an ndarray with axes corresponding to variation with `mu, tau, phi` respectively.
[…]
