+++ docs/index.rst
@@ -5,6 +5,7 @@
 
 Documentation for PythonicDISORT package
 ========================================
+Installation: pip `install PythonicDISORT`
 
 Modules (Basic Documentation)
 =============================

[…]
