### 2024-06-04 fc9206f
Completed changes to address JOSS reviewer feedback.
Fixed error in paper.bib and changed .gitignore.

Finally re-added PyTests as history re-write is complete.

Update README.md

Added Matplotlib to dependencies and corrected minor error in Jupyter Notebook documentation.
