+++ docs/JOSS_Paper/paper.md
@@ -58,13 +58,28 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 --->
 
 # Summary
+<!---
 The Radiative Transfer Equation (RTE) models the processes of absorption, scattering and emission 
-as electromagnetic radiation propagates through a medium. We address the 1D RTE (\ref{RTE}) 
+as electromagnetic radiation propagates through a medium. I address the 1D RTE (\ref{RTE}) 
 in a plane-parallel atmosphere and consider three sources: 
 blackbody emission from the atmosphere $s(\tau)$, scattering from sunlight
 $\frac{\omega I_0}{4 \pi} p\left(\mu, \phi ;-\mu_{0}, \phi_{0}\right) \exp\left(-\mu_{0}^{-1} \tau\right)$,
 and incoming radiation from other atmospheric layers or the Earth's surface modeled by
 Dirichlet boundary conditions.
+--->
+
+The Radiative Transfer Equation (RTE) models the processes of absorption, scattering and emission 
+as electromagnetic radiation propagates through a medium. 
+Consider a plane-parallel, horizontally homogeneous atmosphere with vertical coordinate 
+$\tau$ (optical depth) increasing from top to bottom and directional coordinates $\phi$ for the azimuthal angle (positive is counterclockwise) 
+and $\mu=\cos\theta$ for the polar direction ($\theta$ is the polar angle measured from the surface normal), 
+with $\mu > 0$ pointing up following the convention of [@STWJ1988].
+Given three possible sources: 
+blackbody emission from the atmosphere $s(\tau)$, 
+scattering from a direct beam (of sunlight) with intensity $I_0$ and incident cosine polar and azimuthal angles $\mu_0, \phi_0$,
+and/or radiation from other atmospheric layers or the Earth's surface which is modeled by Dirichlet boundary conditions,
+the diffuse intensity $u(\tau, \mu, \phi)$ propagating in direction $(\mu, \phi)$ 
+is described by the 1D RTE [@Cha1960; @STWJ1988]:
 
 \begin{align}
 \begin{split}
@@ -73,46 +88,65 @@ Dirichlet boundary conditions.
 \end{split} \label{RTE}
 \end{align}
 
-The RTE is important in many fields of science and engineering.
+Here $\omega$ is the single-scattering albedo and $p$ the scattering phase function.
+These are assumed to be independent of $\tau$, i.e. homogeneous in the atmospheric layer.
+An atmosphere with $\tau$-dependent $\omega$ and $p$ can be modelled by 
+a multi-layer atmosphere with different $\omega$ and $p$ for each layer.
+
+The RTE is important in many fields of science and engineering,
+for example, in the retrieval of optical properties from measurements [@TCCGL1999; @MRO/CRISM2008; @TLZWSY2020].
 The gold standard for numerically solving the 1D RTE is the Discrete Ordinate Radiative Transfer 
-package `DISORT` which was coded in FORTRAN 77 and first released in 1988 [@STWJ1988].
+package `DISORT` which was coded in FORTRAN 77 and first released in 1988 [@STWJ1988; @Sta1999].
 It has been widely used, for example by `MODTRAN` [@Ber2014], `Streamer` [@Key1998], and `SBDART` [@Ric1998],
-all of which are comprehensive radiative transfer models that are themselves widely used in atmospheric science.
+all of which are comprehensive radiative transfer models that are themselves widely used in atmospheric science,
+and by the three retrieval papers @TCCGL1999; @MRO/CRISM2008; @TLZWSY2020.
 `DISORT` implements the Discrete Ordinates Method which has two key steps.
-First, the diffuse intensity function $u$ is expanded as the Fourier cosine series:
[…]
