+++ docs/JOSS_Paper/paper.bib
@@ -1,3 +1,125 @@
+@article{TLZWSY2020,
+author = {Teng, Shiwen and Liu, Chao and Zhang, Zhibo and Wang, Yuan and Sohn, Byung-Ju and Yung, Yuk L.},
+title = {Retrieval of Ice-Over-Water Cloud Microphysical and Optical Properties Using Passive Radiometers},
+journal = {Geophysical Research Letters},
+volume = {47},
+number = {16},
+pages = {e2020GL088941},
+keywords = {cloud properties, satellite radiometer, vertical distribution},
+doi = {10.1029/2020GL088941},
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2020GL088941},
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2020GL088941},
+note = {e2020GL088941 2020GL088941},
+abstract = {Abstract Current satellite cloud products from passive radiometers provide effective single-layer cloud properties by assuming a homogeneous cloud in a pixel, resulting in inevitable biases when multiple-layer clouds are present in a vertical column. We devise a novel method to retrieve cloud vertical properties for ice-over-water clouds using passive radiometers. Based on the absorptivity differences of liquid water and ice clouds at four shortwave-infrared channels (centered at 0.87, 1.61, 2.13, and 2.25 μm), cloud optical thicknesses (COT) and effective radii of both upper-layer ice and lower-layer liquid water clouds are inferred simultaneously. The algorithm works most effectively for clouds with ice COT < 7 and liquid water COT > 5. The simulated spectral reflectances based on our retrieved ice-over-water clouds become more consistent with observations than those with a single-layer assumption. This new algorithm will improve our understanding of clouds, and we suggest that these four cloud channels should be all included in future satellite sensors.},
+year = {2020}
+}
+
+@article{TCCGL1999,
+author = {Torricella, Francesca and Cattani, Elsa and Cervino, Marco and Guzzi, Rodolfo and Levoni, Chiara},
+title = {Retrieval of aerosol properties over the ocean using Global Ozone Monitoring Experiment measurements: Method and applications to test cases},
+journal = {Journal of Geophysical Research: Atmospheres},
+volume = {104},
+number = {D10},
+pages = {12085-12098},
+doi = {10.1029/1999JD900040},
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/1999JD900040},
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/1999JD900040},
[…]
