### 2024-06-04 #6442 @dhjx1996
> @dhjx1996 thank you for responding to my suggestions! My last request is just a few more docstrings (_assemble_intensity_and_fluxes.u, _assemble_intensity_and_fluxes.u0, _assemble_intensity_and_fluxes.flux_up, etc.). I know these are internal functions that likely won't be used by users, but it will facilitate easier development later down the road if other people touch the code.
>
> Aside from this, all the changes are to my satisfaction.

Added the docstrings using the same descriptions as in the docstring of `pydisort`.
