### 2024-06-10 b35d8c5
Enabled each function output to switch to its antiderivative by passing the argument `is_antiderivative_wrt_tau = True`. Ready to release v0.8.0.
