+++ src/PythonicDISORT/pydisort.py
@@ -85,22 +85,29 @@ def pydisort(
     Fp(tau) : function, function
         (Energetic) Flux function with argument `tau` (type: array or float) for positive (upward) `mu` values.
         Returns the diffuse flux magnitudes (same type and size as `tau`).
+        Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
+        to switch to an antiderivative of the function with respect to `tau`.
     Fm(tau) : function
         (Energetic) Flux function with argument `tau` (type: array or float) for negative (downward) `mu` values.
         Returns a tuple of the diffuse and direct flux magnitudes respectively where each entry is of the
         same type and size as `tau`.
+        Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
+        to switch to an antiderivative of the function with respect to `tau`.
     u0(tau) : function
         Zeroth Fourier mode of the intensity with argument `tau` (type: array or float).
         Returns an ndarray with axes corresponding to variation with `mu` and `tau` respectively.
         This function is useful for calculating actinic fluxes and other quantities of interest,
         but reclassification of delta-scaled flux and other corrections must be done manually
         (for actinic flux `generate_diff_act_flux_funcs` will automatically perform the reclassification).
+        Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
+        to switch to an antiderivative of the function with respect to `tau`.
     u(tau, phi) : function, optional
-        Intensity function with arguments `(tau, phi, return_Fourier_error=False)` 
-        of types `(array or float, array or float, bool)`.
+        Intensity function with arguments `(tau, phi)` of types `(array or float, array or float)`.
         Returns an ndarray with axes corresponding to variation with `mu, tau, phi` respectively.
-        The optional flag `return_Fourier_error` determines whether the function will also return
-        the Cauchy / Fourier convergence evaluation (type: float) for the last Fourier term.
+        Pass `return_Fourier_error = True` (defaults to `False`) to return the 
+        Cauchy / Fourier convergence evaluation (type: float) for the last Fourier term.
+        Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
+        to switch to an antiderivative of the function with respect to `tau`.
 
     """
     
@@ -313,9 +320,9 @@ def pydisort(
             _autograd_compatible,
         )
         
-        # TMS correction
+        # TMS correction for the intensity (see Section 3.7.2)
         # --------------------------------------------------------------------------------------------------------------------------
-        def TMS_correction(tau, phi):
+        def TMS_correction(tau, phi, is_antiderivative_wrt_tau):
             Ntau = len(tau)
             Nphi = len(phi)
             # Atmospheric layer indices
@@ -360,20 +367,36 @@ def pydisort(
             )
             # --------------------------------------------------------------------------------------------------------------------------
             
-            TMS_correction_pos = (
-                np.exp(-scaled_tau / mu0)[None, :]
-                - np.exp(
-                    (scaled_tau - scaled_tau_arr_l)[None, :] / mu_arr_pos[:, None]
-                    - scaled_tau_arr_l[None, :] / mu0
+            if is_antiderivative_wrt_tau:
+                TMS_correction_pos = (
+                    np.exp(-scaled_tau / mu0)[None, :] / (-scale_tau[None, l] / mu0)
+                    - np.exp(
+                        (scaled_tau - scaled_tau_arr_l)[None, :] / mu_arr_pos[:, None]
+                        - scaled_tau_arr_l[None, :] / mu0
+                    ) / (scale_tau[None, l] / mu_arr_pos[:, None])
                 )
[…]
