### 2024-06-10 676ae37
Updated latest version in JOSS paper.
