### 2024-06-10 8110ec6
Slightly improved documentation on the tau-integration functionality.
