### 2024-08-01 13df1df
Minor correction. src/PythonicDISORT/subroutines.py: return result_without_reclassification + act_dscale_reclassification |-> np.squeeze(result_without_reclassification + act_dscale_reclassification)[()]
