### 2024-08-03 26bf3c7
Temporary commit -- fixed slight bug for autograd compatible isotropic source
