### 2024-08-06 8b04cd8
Fixed major bugs with respect to the isotropic source (Thank you Richard Larsson for catching these bugs!). Added an interpolation subroutine with discussion and accuracy tests in the Jupyter Notebook.
