+++ src/PythonicDISORT/subroutines.py
@@ -1,7 +1,6 @@
 import numpy as np
 import scipy as sc
 from math import pi
-from numpy.polynomial.legendre import leggauss
 
 
 def transform_interval(arr, c, d, a, b):
@@ -71,7 +70,7 @@ def calculate_nu(mu, phi, mu_p, phi_p):
 
     Returns
     -------
-    ndarray
+    nu : ndarray
         Cosine of scattering angles which axes capture variation with `mu, phi, mu_p, phi_p` respectively.
 
     """
@@ -104,7 +103,7 @@ def Gauss_Legendre_quad(N, c=0, d=1):
         Quadrature weights.
 
     """
-    mu_arr_pos, W = leggauss(int(N))
+    mu_arr_pos, W = np.polynomial.legendre.leggauss(int(N))
 
     return transform_interval(mu_arr_pos, c, d, -1, 1), transform_weights(W, c, d, -1, 1)
 
@@ -224,8 +223,10 @@ def atleast_2d_append(*arys):
         
 def generate_diff_act_flux_funcs(u0):
     """Generates the up and down diffuse actinic flux functions respectively.
-    This is a use case of the ``u0`` function that is an output of ``pydisort``.
+    This is a use case of the ``u0`` function which is an output of ``pydisort``.
     The reclassification of delta-scaled actinic flux is automatically performed.
+    The computation of actinic fluxes is similar to that of energetic fluxes
+    and the latter is discussed in section 3.8 of the Comprehensive Documentation.
 
     Parameters
     ----------
@@ -235,12 +236,12 @@ def generate_diff_act_flux_funcs(u0):
 
     Returns
     -------
-    function
+    Fp_act(tau) : function
         Actinic flux function with argument `tau` (type: array) for positive (upward) `mu` values.
         Returns the diffuse flux magnitudes (type: array).
         Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
         to switch to an antiderivative of the function with respect to `tau`.
-    function
+    Fm_act(tau) : function
         Actinic flux function with argument `tau` (type: array) for negative (downward) `mu` values.
         Returns the diffuse flux magnitudes (type: array).
         Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
@@ -262,6 +263,65 @@ def generate_diff_act_flux_funcs(u0):
     return flux_act_up, flux_act_down_diffuse
 
 
+def interpolate(mu_arr_pos, u):
+    """Polynomial (Barycentric) interpolation with respect to mu. The output 
+    is a function that is continuous and variable in all three arguments: mu, tau and phi.
+    Discussed in sections 3.7 and 6.3 in the Comprehensive Documentation.
+
+    Parameters
+    ----------
+    mu_arr_pos : array
+        Positive `mu` (cosine of polar angle) quadrature nodes.
+        Equivalent to `mu_arr[:N]`.
+    u : function
+        Non-interpolated intensity function as outputted by `pydisort`.
+
+    Returns
+    -------
+    u_interpol : function
+        Intensity function with arguments `(mu, tau, phi)` each of type array or float.
+        Returns an ndarray with axes corresponding to variation with `mu, tau, phi` respectively.
+        Pass `return_Fourier_error = True` (defaults to `False`) to return the 
+        Cauchy / Fourier convergence evaluation (type: float) for the last Fourier term.
+        Pass `is_antiderivative_wrt_tau = True` (defaults to `False`)
+        to switch to an antiderivative of the function with respect to `tau`.
+
+    """
+    N = len(mu_arr_pos)
[…]
