### 2024-08-06 d80de76
Added comments to the initial setup part of the Jupyter Notebook of each pytest to faciliate their role as examples of use.
