+++ .github/workflows/continuous_distribution.yml
@@ -46,7 +46,7 @@ jobs:
     if: "!github.event.release.prerelease"
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/download-artifact@v4.1.7
+      - uses: actions/download-artifact@v4.1.8
         with:
           name: releases
           path: dist
