### 2024-09-06 397323c
Completed optimizations suggested by Richard Larsson. Corrected a bug in the `is_antiderivative_wrt_tau` feature of the multilayer TMS correction. Need to do a bit of documentation clean-up then version 0.9.1 can be released.
