+++ src/PythonicDISORT/_solve_for_coeffs.py
@@ -1,5 +1,5 @@
 from PythonicDISORT.subroutines import _mathscr_v
-from PythonicDISORT.subroutines import _nd_slice_to_indexes
+from PythonicDISORT.subroutines import to_diag_ordered_form
 
 import numpy as np
 import scipy as sc
@@ -23,10 +23,10 @@ def _solve_for_coeffs(
     there_is_beam_source,                   # Is there a beam source?
     b_pos, b_neg,                           # Dirichlet BCs
     b_pos_is_scalar, b_neg_is_scalar,       # Is each Dirichlet BCs scalar (isotropic)?
-    s_poly_coeffs,                          # Polynomial coefficients of isotropic source           
     Nscoeffs,                               # Number of isotropic source polynomial coefficients
+    s_poly_coeffs,                          # Polynomial coefficients of isotropic source           
     there_is_iso_source,                    # Is there an isotropic source?
-    NLayers_to_use_sparse,                  # Number of layers above or equal which to use sparse matrices
+    NLayers_for_banded_solver,              # Number of layers above or equal which to use `scipy.linalg.solve_banded`
 ):
     """
     Uses the boundary conditions to solve for the unknown coefficients 
@@ -62,11 +62,11 @@ def _solve_for_coeffs(
     | `b_pos`                        | `NQuad/2 x NFourier` or scalar         |
     | `b_neg`                        | `NQuad/2 x NFourier` or scalar         |
     | `b_pos_is_scalar`              | boolean                                |
-    | `b_neg_is_scalar`              | boolean                                |              
-    | `s_poly_coeffs`                | `NLayers x Nscoeffs` or `Nscoeffs`     |
+    | `b_neg_is_scalar`              | boolean                                |
     | `Nscoeffs`                     | scalar                                 |
+    | `s_poly_coeffs`                | `NLayers x Nscoeffs` or `Nscoeffs`     |
     | `there_is_iso_source`          | boolean                                |
-    | `NLayers_to_use_sparse`        | scalar                                 |
+    | `NLayers_for_banded_solver`    | scalar                                 |
     
     Notable internal variables of _solve_for_coeffs
     |   Variable   |                 Shape                |
@@ -77,13 +77,16 @@ def _solve_for_coeffs(
     ################################## Solve for coefficients of homogeneous solution ##########################################
     ############# Refer to Section 3.6.2 (single-layer) and 4 (multi-layer) of the Comprehensive Documentation #################
     
+    dim = NLayers * NQuad
     GC_collect = np.empty((NFourier, NLayers, NQuad, NQuad))
-    use_sparse_framework = NLayers >= NLayers_to_use_sparse
+    use_banded_solver = (NLayers >= NLayers_for_banded_solver)
+    if use_banded_solver:
+        sym_offset = 3 * N - 1
     
     # The following loops can easily be parallelized, but the speed-up is unlikely to be worth the overhead
     for m in range(NFourier):
         m_equals_0 = (m == 0)
-        BDRF_bool = m < NBDRF
+        BDRF_bool = (NBDRF > m) 
         
         G_collect_m = G_collect[m, :, :, :]
         K_collect_m = K_collect[m, :, :]
@@ -122,12 +125,12 @@ def _solve_for_coeffs(
             b_neg_m = b_neg[:, m]
         
         # _mathscr_v_contribution
[…]
