+++ src/PythonicDISORT/pydisort.py
@@ -22,7 +22,7 @@ def pydisort(
     NT_cor=False,
     BDRF_Fourier_modes=[],
     s_poly_coeffs=np.array([[]]),
-    NLayers_to_use_sparse=13,
+    NLayers_for_banded_solver=13,
     _autograd_compatible = False,
 ):
     """Solves the 1D RTE for the fluxes, and optionally intensity,
@@ -75,8 +75,8 @@ def pydisort(
     s_poly_coeffs : optional, array
         Polynomial coefficients of isotropic internal sources.
         Arrange coefficients from lowest order term to highest.
-    NLayers_to_use_sparse : optional, int
-        At or above how many atmospheric layers should SciPy's sparse matrix framework be used?
+    NLayers_for_banded_solver : optional, int
+        At or above how many atmospheric layers should `scipy.linalg.solve_banded` be used?
 
     Returns
     -------
@@ -115,25 +115,25 @@ def pydisort(
     
     """
     Arguments of pydisort
-    |         Variable        |            Type / Shape            |
-    | ----------------------- | ---------------------------------- |
-    | `tau_arr`               | `NLayers`                          |
-    | `omega_arr`             | `NLayers`                          |
-    | `NQuad`                 | scalar                             |
-    | `Leg_coeffs_all`        | `NLeg_all`                         |
-    | `mu0`                   | scalar                             |
-    | `I0`                    | scalar                             |
-    | `phi0`                  | scalar                             |
-    | `NLeg`                  | scalar                             |
-    | `NFourier`              | scalar                             |
-    | `b_pos`                 | `NQuad/2 x NFourier` or scalar     |
-    | `b_neg`                 | `NQuad/2 x NFourier` or scalar     |              
-    | `only_flux`             | boolean                            |
-    | `f_arr`                 | `NLayers` or scalar                |
-    | `NT_cor`                | boolean                            |
-    | `BDRF_Fourier_modes`    | `NBDRF`                            |
-    | `s_poly_coeffs`         | `NLayers x Nscoeffs` or `Nscoeffs` |
-    | `NLayers_to_use_sparse` | scalar                             |
+    |          Variable           |            Type / Shape            |
+    | --------------------------- | ---------------------------------- |
+    | `tau_arr`                   | `NLayers`                          |
+    | `omega_arr`                 | `NLayers`                          |
+    | `NQuad`                     | scalar                             |
+    | `Leg_coeffs_all`            | `NLeg_all`                         |
+    | `mu0`                       | scalar                             |
+    | `I0`                        | scalar                             |
+    | `phi0`                      | scalar                             |
+    | `NLeg`                      | scalar                             |
+    | `NFourier`                  | scalar                             |
+    | `b_pos`                     | `NQuad/2 x NFourier` or scalar     |
+    | `b_neg`                     | `NQuad/2 x NFourier` or scalar     |              
+    | `only_flux`                 | boolean                            |
+    | `f_arr`                     | `NLayers` or scalar                |
+    | `NT_cor`                    | boolean                            |
+    | `BDRF_Fourier_modes`        | `NBDRF`                            |
+    | `s_poly_coeffs`             | `NLayers x Nscoeffs` or `Nscoeffs` |
+    | `NLayers_for_banded_solver` | scalar                             |
     
[…]

[…]
