### 2024-09-19 8adfb3c
Made the optional argument `autograd_compatible` visible (it was previously `_autograd_compatible`).
