+++ pydisotest/11_test.py
@@ -31,6 +31,7 @@ def test_11a():
     NLoops=None
     only_flux=False
     use_banded_solver_NLayers=10
+	autograd_compatible=False
 
     ####################################################################################################
     
