+++ pydisotest/4_test.py
@@ -53,6 +53,7 @@ def test_4a():
     BDRF_Fourier_modes=[]
     s_poly_coeffs=np.array([[]])
     use_banded_solver_NLayers=10
+	autograd_compatible=False
 
     ####################################################################################################
 
@@ -128,6 +129,7 @@ def test_4b():
     BDRF_Fourier_modes=[]
     s_poly_coeffs=np.array([[]])
     use_banded_solver_NLayers=10
+	autograd_compatible=False
 
     ####################################################################################################
 
@@ -203,6 +205,7 @@ def test_4c():
     BDRF_Fourier_modes=[]
     s_poly_coeffs=np.array([[]])
     use_banded_solver_NLayers=10
+	autograd_compatible=False
 
     ####################################################################################################
 
