### 2024-10-04 #6442 @pscicluna
I'm very sorry this took rather a while to get around to. This paper and repo seems to mostly be in a very good state, almost ready to go! I do still have a few further (rather minor) suggestions for improvements, however, so I will keep this review very short.

My biggest concern is about the examples, as for the other reviewers. The docs are great for explaining *what* PythonicDISORT does, but the examples of *when* to use it are focused on technical details. I would appreciate some context for what they mean, and what sort of problem that might mean I apply the code to. For example, `3_test.ipynb` tests HG scattering - it would be handy for users if you could explain that this is a useful approximation for a simple case of anisotropic scattering, and describe a simple (real-world) situation where it might be applicable - e.g. scattering by particles small compared to the wavelength, but not in the Rayleigh regime (e.g. dust, aerosols). Similarly, `9_test.ipynb`'s test problem is described as "General Emitting/Absorbing/Scattering Medium with Every Computational Layer Different" - it would be great if you could describe this in terms of one of the (many) problems that fall into this category. I would choose one that you are familiar with. If all the examples had some context like this it would help a new user but things together, especially when they may use different terminology in their field. It would also be really helpful if these tests/examples were part of the docs, but I realise that you do point to them in the main docs.

I am a little concerned about the use of assert for argument validation in e.g. `pydisort`. This makes the code a little difficult to parse, and assertions can be overridden by invoking `python -O`. If these tests are really critical (i.e. they control output correctness or ensure the code doesn't crash) it would be good to see these replaced with some sort of validation library. If they're not so critical (e.g. if they're meant to be "executable comments"), then it's not such a big deal. The same is true of a few of the other routines in the package.
