### 2024-10-08 e78a111
Fixed bug in direct beam multilayer solver. Changed assert-based input checks to if-statements that raise ValueError. Added return_tau_arr argument to every output function. Improvement documentation. Added explanatory preamble to each PyTest.
