+++ docs/JOSS_Paper/paper.bib
@@ -1,3 +1,19 @@
+@article{HP2024
+author = {Ho, Dion J. X. and Pincus, Robert},
+title = {Two-Streams Revisited: General Equations, Exact Coefficients, and Optimized Closures},
+journal = {Journal of Advances in Modeling Earth Systems},
+volume = {16},
+number = {10},
+pages = {e2024MS004504},
+keywords = {two-stream equations, radiative transfer, atmospheric fluxes, optimization, two-stream closures, plane-parallel atmosphere},
+doi = {https://doi.org/10.1029/2024MS004504},
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2024MS004504},
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2024MS004504},
+note = {e2024MS004504 2024MS004504},
+abstract = {Abstract Two-Stream Equations are the most parsimonious general models for radiative flux transfer with one equation to model each of upward and downward fluxes; these are coupled due to the transfer of fluxes between hemispheres. Standard two-stream approximation of the Radiative Transfer Equation assumes that the ratios of flux transferred (coupling coefficients) are both invariant with optical depth and symmetric with respect to upwelling and downwelling radiation. Two-stream closures are derived by making additional assumptions about the angular distribution of the intensity field, but none currently works well for all parts of the optical parameter space. We determine the exact values of the two-stream coupling coefficients from multi-stream numerical solutions to the Radiative Transfer Equation for shortwave radiation. The resulting unique coefficients accurately reconstruct entire flux profiles but depend on optical depth. More importantly, they generally take on unphysical values when symmetry is assumed. We derive a general form of the Two-Stream Equations for which the four coupling coefficients are guaranteed to be physically explicable. While non-constant coupling coefficients are required to reconstruct entire flux profiles, numerically optimized constant coupling coefficients (which admit analytic solutions) reproduced shortwave reflectance and transmittance with relative errors no greater than 4×10−5 \$4\times 1{0}^{-5}\$ over a large range of optical parameters. The optimized coefficients show a dependence on solar zenith angle and total optical depth that diminishes as the latter increases. This explains why existing coupling coefficients, which often omit the former and mostly neglect the latter, tend to work well for only thin or only thick atmospheres.},
+year = {2024}
+}
+
 @article{TLZWSY2020,
 author = {Teng, Shiwen and Liu, Chao and Zhang, Zhibo and Wang, Yuan and Sohn, Byung-Ju and Yung, Yuk L.},
 title = {Retrieval of Ice-Over-Water Cloud Microphysical and Optical Properties Using Passive Radiometers},
