+++ src/PythonicDISORT/subroutines.py
@@ -28,6 +28,7 @@ def transform_interval(arr, c, d, a, b):
     return (((arr - a) * (d - c)) / (b - a)) + c
 
 
+
 def transform_weights(weights, c, d, a, b):
     """Transforms an array of quadrature weights from interval [a, b] to [c, d].
 
@@ -53,9 +54,10 @@ def transform_weights(weights, c, d, a, b):
     return weights * (d - c) / (b - a)
 
 
+
 def calculate_nu(mu, phi, mu_p, phi_p):
-    """Calculates the cosine of the scattering angle `nu` 
-    between incident angle `(mu_p, phi_p)` and scattering angle `(mu, phi)`.
+    """Calculates the cosine of the scattering angle ``nu`` 
+    between incident angle ``(mu_p, phi_p)`` and scattering angle ``(mu, phi)``.
 
     Parameters
     ----------
@@ -71,7 +73,7 @@ def calculate_nu(mu, phi, mu_p, phi_p):
     Returns
     -------
     nu : ndarray
-        Cosine of scattering angles which axes capture variation with `mu, phi, mu_p, phi_p` respectively.
+        Cosine of scattering angles which axes capture variation with ``mu, phi, mu_p, phi_p`` respectively.
 
     """
     mu, phi, mu_p, phi_p = np.atleast_1d(mu, phi, mu_p, phi_p)
@@ -83,12 +85,13 @@ def calculate_nu(mu, phi, mu_p, phi_p):
     return np.squeeze(nu)
 
 
+
 def Gauss_Legendre_quad(N, c=0, d=1):
     """Generates Gauss-Legendre quadrature zero points and weights for integration from c to d.
 
     Parameters
     ----------
-    N : int
+    N : int (will be converted to int)
         Number of quadrature nodes.
     c : float, optional
         Start of integration interval.
@@ -106,6 +109,7 @@ def Gauss_Legendre_quad(N, c=0, d=1):
     mu_arr_pos, W = np.polynomial.legendre.leggauss(int(N))
 
     return transform_interval(mu_arr_pos, c, d, -1, 1), transform_weights(W, c, d, -1, 1)
+
 
 
 def Clenshaw_Curtis_quad(Nphi, c=0, d=(2 * pi)):
@@ -129,8 +133,8 @@ def Clenshaw_Curtis_quad(Nphi, c=0, d=(2 * pi)):
 
     """
     # Ensure that the number of nodes is odd and greater than 2
-    assert Nphi > 2
-    assert Nphi % 2 == 1
+    if not (Nphi > 2 and Nphi % 2 == 1):
+        raise ValueError("The number of quadrature nodes must be odd and greater than 2.")
 
     Nphi -= 1  # The extra index corresponds to the point 0 which we will add later
     Nphi_pos = Nphi // 2
@@ -144,9 +148,10 @@ def Clenshaw_Curtis_quad(Nphi, c=0, d=(2 * pi)):
     return transform_interval(phi_arr, c, d, -1, 1), transform_weights(full_weights_phi, c, d, -1, 1)
 
 
+
 def generate_FD_mat(Ntau, a, b):
     """Generates a sparse first derivative central difference (second-order accuracy) 
-    matrix on [a,b] in `csr` format with `Ntau` grid points.
+    matrix on [a,b] in ``csr`` format with ``Ntau`` grid points.
     Second order forward or backward differences are used at the boundaries.
 
     Parameters
@@ -182,6 +187,7 @@ def generate_FD_mat(Ntau, a, b):
 
     return tau_arr, first_deriv.tocsr()
   
+
 
 def atleast_2d_append(*arys):
     """View inputs as arrays with at least two dimensions. 
@@ -219,7 +225,8 @@ def atleast_2d_append(*arys):
[…]
