### 2024-10-13 c9690d6
Corrected "docs/JOSS_Paper/paper.bib" to ensure JOSS paper compiles.
