+++ docs/JOSS_Paper/paper.bib
@@ -1,4 +1,4 @@
-@article{HP2024
+@article{HP2024,
 author = {Ho, Dion J. X. and Pincus, Robert},
 title = {Two-Streams Revisited: General Equations, Exact Coefficients, and Optimized Closures},
 journal = {Journal of Advances in Modeling Earth Systems},
