### 2024-10-13 #6442 @dhjx1996
@pscicluna Thank you very much for the review!

> I would appreciate some context for what they mean, and what sort of problem that might mean I apply the code to.

I have added a preamble to every test problem Jupyter Notebook, e.g. at the start of `3_test.py`: "The *Henyey-Greenstein (HG) phase function* is the most commonly used approximation for anisotropic phase functions and is appropriate, though oversimple, for clouds and aerosols". In `4_test.py` and `5_test.py` I also added a plot to compare the haze-L and cloud C1 phase functions respectively against a HG phase function of the same asymmetry parameter. I did not add these examples to the main docs so as to not further bloat it, but I do point to them in the main docs and in the readme.

> I am a little concerned about the use of assert for argument validation in e.g. pydisort.

I have replaced all assert-statements with if-statements that raise a ValueError with an error message upon violation. Would this be satisfactory? The input restrictions are all explained in Section 1 of the main docs alongside the explanation of each parameter.
