### 2024-10-13 53831ce
v0.9.3 is identical to v0.9.2 except that two sets of values that were used for debugging are no longer being returned.
