### 2024-10-13 6c0c0e6
Fixed "\docs\JOSS_Paper\paper.bib".
