+++ docs/JOSS_Paper/paper.bib
@@ -6,7 +6,7 @@ volume = {16},
 number = {10},
 pages = {e2024MS004504},
 keywords = {two-stream equations, radiative transfer, atmospheric fluxes, optimization, two-stream closures, plane-parallel atmosphere},
-doi = {https://doi.org/10.1029/2024MS004504},
+doi = {10.1029/2024MS004504},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2024MS004504},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2024MS004504},
 note = {e2024MS004504 2024MS004504},
