### 2024-10-15 #6442 @phibeck
Thank you @pscicluna for working through your checklist, and thanks @dhjx1996 for addressing the issues.

@pscicluna could you please let us know whether this resolves your concerns?
