### 2024-10-16 #6442 @pscicluna
>
> I have added a preamble to every test problem Jupyter Notebook, e.g. at the start of `3_test.py`: "The _Henyey-Greenstein (HG) phase function_ is the most commonly used approximation for anisotropic phase functions and is appropriate, though oversimple, for clouds and aerosols". In `4_test.py` and `5_test.py` I also added a plot to compare the haze-L and cloud C1 phase functions respectively against a HG phase function of the same asymmetry parameter. I did not add these examples to the main docs so as to not further bloat it, but I do point to them in the main docs and in the readme.

I particularly like the additions to the test notebooks. I think this is a good medium between adding them all to the docs and not having any of this sort of documentation.

>
> I have replaced all assert-statements with if-statements that raise a ValueError with an error message upon violation. Would this be satisfactory? The input restrictions are all explained in Section 1 of the main docs alongside the explanation of each parameter.

This is good - now the code will raise an informative error message and the checks can't be avoided. For the future, it might be good to look into a validation library, but this shouldn't stand in the way of acceptance.

I think this resolves all my concerns, thank you!
