### 2024-10-17 cbb393c
Changed the default of `use_banded_solver_NLayers` from 10 to 13. The old value is a carry-over from older versions. Two new subtests: "3multilayer" and "8ARTS" have been added to better test the multilayer solver.
