+++ pydisotest/3_test.py
@@ -157,3 +157,79 @@ def test_3b():
     assert np.max(ratio_flux_down_direct[diff_flux_down_direct > 1e-3], initial=0) < 1e-3
     assert np.max(diff_ratio[diff > 1e-3], initial=0) < 1e-2
     # --------------------------------------------------------------------------------------------------
+    
+
+def test_3multilayer():
+    print()
+    print("################################################ Test 3multilayer ##################################################")
+    print()
+    ######################################### PYDISORT ARGUMENTS #######################################
+
+    tau_arr = np.cumsum(0.5 * np.arange(7))[1:]
+    omega_arr = np.linspace(1 - 1e-6, 0.8, 6)  # Reduced from 1 (conservative scattering) which is not implemented
+    NQuad = 16
+    Leg_coeffs_all = np.linspace(0.9, 0.65, 6)[:, None] ** np.arange(32)[None, :]
+    mu0 = 1
+    I0 = pi / mu0
+    phi0 = pi
+
+    # Optional (used)
+    f_arr = Leg_coeffs_all[:, NQuad]
+    NT_cor = True
+
+    # Optional (unused)
+    NLeg=None
+    NLoops=None
+    b_pos=0
+    b_neg=0
+    only_flux=False
+    BDRF_Fourier_modes=[]
+    s_poly_coeffs=np.array([[]])
+    use_banded_solver_NLayers=10
+    autograd_compatible=False
+
+    ####################################################################################################
+
+    # Call pydisort function
+    mu_arr, flux_up, flux_down, u0, u = PythonicDISORT.pydisort(
+        tau_arr, omega_arr,
+        NQuad,
+        Leg_coeffs_all,
+        mu0, I0, phi0,
+        f_arr=f_arr,
+        NT_cor=NT_cor
+    )
+    
+    # mu_arr is arranged as it is for code efficiency and readability
+    # For presentation purposes we re-arrange mu_arr from smallest to largest
+    reorder_mu = np.argsort(mu_arr)
+    mu_arr_RO = mu_arr[reorder_mu]
+
+    # We may not want to compare intensities around the direct beam
+    deg_around_beam_to_not_compare = 0
+    mu_to_compare = (
+        np.abs(np.arccos(np.abs(mu_arr_RO)) - np.arccos(mu0)) * 180 / pi
+        > deg_around_beam_to_not_compare
+    )
+
+    
+    # Load results from version 4.0.99 of Stamnes' DISORT for comparison
+    results = np.load("Stamnes_results/3multilayer_test.npz")
+    
+    # Perform the comparisons
+    (
+        diff_flux_up,
+        ratio_flux_up,
+        diff_flux_down_diffuse,
+        ratio_flux_down_diffuse,
+        diff_flux_down_direct,
+        ratio_flux_down_direct,
+        diff,
+        diff_ratio,
+    ) = _compare(results, mu_to_compare, reorder_mu, flux_up, flux_down, u)
+    
+    assert np.max(ratio_flux_up[diff_flux_up > 1e-3], initial=0) < 1e-3
+    assert np.max(ratio_flux_down_diffuse[diff_flux_down_diffuse > 1e-3], initial=0) < 1e-3
+    assert np.max(ratio_flux_down_direct[diff_flux_down_direct > 1e-3], initial=0) < 1e-3
+    assert np.max(diff_ratio[diff > 1e-3], initial=0) < 1e-2
+    # --------------------------------------------------------------------------------------------------
