+++ pydisotest/inpydis.py
@@ -1,0 +1,2330 @@
+import numpy as np
+
+N = 20
+
+src = [[]]*101
+tau = [[]]*101
+
+src[0] = np.array([
+[1.2861096791103891e-15, -7.929732934312494e-17],
+[1.2737267352367948e-15, -7.158789313540497e-17],
+[1.2097873061017914e-15, -5.393181653700997e-17],
+[8.432180169549867e-16, 4.2884515490693675e-18],
+[6.70376675373952e-16, 2.2820163572842812e-17],
+[7.153500080114476e-16, 1.9108697160885224e-17],
+[6.868685024971062e-16, 2.1061015597436515e-17],
+[6.305561702879067e-16, 2.4410930679657766e-17],
+[5.079445814986021e-16, 3.095213600381241e-17],
+[6.408934552160232e-16, 2.4388741479783465e-17],
+[2.4673473199233786e-15, -6.17024982540027e-17],
+[6.063567794057963e-15, -2.2748221379547097e-16],
+[2.4107740741361433e-14, -1.0518341101689843e-15],
+[6.872366231870501e-14, -3.0835846621865728e-15],
+[2.2350773645368494e-13, -1.0124545441441945e-14],
+[8.128112498077814e-13, -3.692240817253281e-14],
+[1.540659381697786e-12, -7.001741766989075e-14],
+[1.146055130602766e-13, -5.1766329796140815e-15],
+[-5.794331641115285e-11, 2.634630301317078e-12],
+[-3.9033272428780745e-10, 1.774787162724635e-11]
+]).reshape(N, 2);
+tau[0] = np.array([1.6062061530769762, 3.6213837643191846, 6.296248732569003, 9.326787582243066, 12.117402569669133, 14.588555320234043, 16.81007751750537, 18.744494739606758, 20.25611491594599, 21.21532771923333, 21.692765380786426, 21.888920285964463, 21.959350045722218, 21.98337400074956, 21.990690797530917, 21.99268539107416, 21.993161795454895, 21.993245479401374, 21.993257482521997, 21.99325904468834])
+src[1] = np.array([
+[1.2861105414222943e-15, -7.929738507133146e-17],
+[1.2737275622369953e-15, -7.158792665469366e-17],
+[1.2097879435571723e-15, -5.393179950639235e-17],
+[8.432187025345697e-16, 4.288433067778769e-18],
+[6.703824275169003e-16, 2.2819565870913322e-17],
+[7.153800660727447e-16, 1.910612211786954e-17],
+[6.870467891676557e-16, 2.1048222993647973e-17],
+[6.314935671157024e-16, 2.435263023580597e-17],
+[5.120532371565539e-16, 3.072247318895945e-17],
+[6.486270725602736e-16, 2.3986012897124008e-17],
+[2.433756077512331e-15, -6.00234321349867e-17],
+[5.902860961535496e-15, -2.1959631356701554e-16],
+[2.3284147036699205e-14, -1.0117020320910082e-15],
+[6.62282103853879e-14, -2.9622541574401227e-15],
+[2.1523145126825785e-13, -9.722416314977033e-15],
+[7.825556972921149e-13, -3.5452585346225496e-14],
+[1.4832712501605477e-12, -6.722953001075277e-14],
+[1.1036281915026586e-13, -4.970524257315825e-15],
+[-5.5783890601411e-11, 2.5297264832852243e-12],
+[-3.757824199212579e-10, 1.7041036774480826e-11]
+]).reshape(N, 2);
[…]
