+++ docs/JOSS_Paper/paper.md
@@ -75,9 +75,9 @@ $\tau$ (optical depth) increasing from top to bottom and directional coordinates
 and $\mu=\cos\theta$ for the polar direction ($\theta$ is the polar angle measured from the surface normal), 
 with $\mu > 0$ pointing up following the convention of [@STWJ1988].
 Given three possible sources: 
-blackbody emission from the atmosphere $s(\tau)$, 
-scattering from a collimated beam of starlight with intensity $I_0$ and incident cosine polar and azimuthal angles $\mu_0, \phi_0$,
-and/or radiation from other atmospheric layers or the Earth's surface which is modeled by Dirichlet boundary conditions,
+blackbody emission from the atmosphere $s(\tau)$; 
+scattering from a collimated beam of starlight with intensity $I_0$ and incident azimuthal and cosine polar angles $\phi_0, \mu_0$;
+radiation from other atmospheric layers or the Earth's surface which is modeled by Dirichlet boundary conditions,
 the diffuse intensity $u(\tau, \mu, \phi)$ propagating in direction $(\mu, \phi)$ 
 is described by the 1D RTE [@Cha1960; @STWJ1988]:
 
@@ -90,11 +90,11 @@ is described by the 1D RTE [@Cha1960; @STWJ1988]:
 
 Here $\omega$ is the single-scattering albedo and $p$ the scattering phase function.
 These are assumed to be independent of $\tau$, i.e. homogeneous in the atmospheric layer.
-An atmosphere with $\tau$-dependent $\omega$ and $p$ can be modelled by 
+An atmosphere with $\tau$-dependent $\omega$ and $p$ can be modeled by 
 a multi-layer atmosphere with different $\omega$ and $p$ for each layer.
 
 The RTE is important in many fields of science and engineering,
-for example, in the retrieval of optical properties from measurements [@TCCGL1999; @MRO/CRISM2008; @TLZWSY2020].
+for example, in the retrieval of optical properties of the medium from measurements [@TCCGL1999; @MRO/CRISM2008; @TLZWSY2020].
 The gold standard for numerically solving the 1D RTE is the Discrete Ordinate Radiative Transfer 
 package `DISORT` which was coded in FORTRAN 77 and first released in 1988 [@STWJ1988; @Sta1999].
 It has been widely used, for example by `MODTRAN` [@Ber2014], `Streamer` [@Key1998], and `SBDART` [@Ric1998],
@@ -114,23 +114,25 @@ where $\gamma$ is the scattering angle.
 These address the $\phi'$ integral in (\ref{RTE}) and decompose the problem into solving
 
 $$
-\mu \frac{d u^m(\tau, \mu)}{d \tau}=u^m(\tau, \mu)-\int_{-1}^1 D^m\left(\mu, \mu'\right) u^m\left(\tau, \mu'\right) \mathrm{d} \mu' - Q^m(\tau, \mu) - \delta_{0m}s(\tau)
+\mu \frac{\partial u^m(\tau, \mu)}{\partial \tau}=u^m(\tau, \mu)-\int_{-1}^1 D^m\left(\mu, \mu'\right) u^m\left(\tau, \mu'\right) \mathrm{d} \mu' - Q^m(\tau, \mu) - \delta_{0m}s(\tau)
 $$
 
 for each Fourier mode of $u$. The terms $D^m$ are derived from $p$ 
 and are thus also independent of $\tau$. The second key step is to discretize the 
-$\mu'$ integral using some quadrature scheme; `DISORT` 
+$\mu'$ integral using some quadrature scheme. `DISORT` 
 uses the double-Gauss quadrature scheme from @Syk1951. 
-This results in a system of ordinary differential equations that can be solved using standard methods.
[…]
