+++ docs/JOSS_Paper/paper.bib
@@ -1,3 +1,31 @@
+@article{Wis1977,
+      author = "W. J.  Wiscombe",
+      title = "The Delta–M Method: Rapid Yet Accurate Radiative Flux Calculations for Strongly Asymmetric Phase Functions",
+      journal = "Journal of Atmospheric Sciences",
+      year = "1977",
+      publisher = "American Meteorological Society",
+      address = "Boston MA, USA",
+      volume = "34",
+      number = "9",
+      doi = "10.1175/1520-0469(1977)034<1408:TDMRYA>2.0.CO;2",
+      pages=      "1408 - 1422",
+      url = "https://journals.ametsoc.org/view/journals/atsc/34/9/1520-0469_1977_034_1408_tdmrya_2_0_co_2.xml"
+}
+
+@article{NT1988,
+title = {Algorithms for radiative intensity calculations in moderately thick atmospheres using a truncation approximation},
+journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+volume = {40},
+number = {1},
+pages = {51-69},
+year = {1988},
+issn = {0022-4073},
+doi = {10.1016/0022-4073(88)90031-3},
+url = {https://www.sciencedirect.com/science/article/pii/0022407388900313},
+author = {T. Nakajima and M. Tanaka},
+abstract = {The efficiency of numerical calculations is discussed for selected algorithms employing the discrete ordinate method and the truncation approximation for the solar radiative intensity in moderately thick, plane-parallel scattering atmospheres. It is found that truncation of the phase function causes a significant error in the computed intensity and the magnitude of this error depends significantly on how the intensity is retrieved from the truncated radiative transfer equation. A newly developed retrieval algorithm, the IMS- method, yields the intensity field with an error ⪅1% when the number of discrete path is as small as 10 in the hemisphere for aerosol-laden atmospheres with optical thickness ⪅1.}
+}
+
 @article{HP2024,
 author = {Ho, Dion J. X. and Pincus, Robert},
 title = {Two-Streams Revisited: General Equations, Exact Coefficients, and Optimized Closures},
