### 2024-10-29 #6442 @dhjx1996
> Thanks @dhjx1996 could you please make sure the version tag and title are the same for archived version, and post the version number here? Please also confirm again that you checked the remaining points on the checklist, i.e. authors and affiliations, title and author list, and the license.

The Zenodo archive title is _PythonicDISORT: A Python reimplementation of the Discrete Ordinate Radiative Transfer package DISORT (v0.9.3; JOSS-published release)_ the paper title is _PythonicDISORT: A Python reimplementation of the Discrete Ordinate Radiative Transfer package DISORT_.

Version number: 0.9.3

I have also checked the author list (including ORCIDs) and license.
