### 2024-10-29 #6442 @dhjx1996
> Hi @dhjx1996 thanks for confirming. Sorry I wasn't being clear. Could you please adjust the version tag in Zenodo to match that of the repo, i.e. v0.9.3, and remove the add-on "(v0.9.3; JOSS-published release)" in the title so that it matches the JOSS paper? Thank you

Hi @phibeck no problem! I have made the change. The Zenodo version tag is now 0.9.3 and the title is exactly _PythonicDISORT: A Python reimplementation of the Discrete Ordinate Radiative Transfer package DISORT_ as per the JOSS paper.
