### 2024-10-29 #6442 @phibeck
Thanks, if you could do one last adjustment, please unify the version number (version 0.9.3 on zenodo vs version v0.9.3 in the repo), thank you! I will then go through my checklist and offer comments on the manuscript.
