+++ .github/workflows/continuous_distribution.yml
@@ -51,7 +51,7 @@ jobs:
           name: releases
           path: dist
       - name: Publish package to PyPI
-        uses: pypa/gh-action-pypi-publish@v1.10.3
+        uses: pypa/gh-action-pypi-publish@v1.12.0
         with:
           user: __token__
           password: ${{ secrets.PYPI_TOKEN }}
