### 2024-11-07 75348b6
Fixed bug in u0. Changed PythonicDISORT.subroutines.interpolate to now be able to interpolate u0.
