### 2024-11-08 #6442 @dhjx1996
> Hi @dhjx1996, thanks for your patience. Here are a few more comments/suggestions for the manuscript. Please have a look when you have a moment.
>
> * line 4: in your affiliations, please spell out USA as United States of America
> * line 9f: I would recommend "...from top to bottom, directional coordinates... , and \mu = cos\theta..
> * line 11: please omit [] brackets around the reference since it is object of the sentence
> * line 12: please change to: "Given three possible sources, namely blackbody emission.., scattering from ..., and radiation from other ...., the diffuse intensity....
> * line 14: "incident azimuthal and cosine polar angles \phi_0 and \mu_0, respectively"
> * line 24: "which was written in FORTRAN77"
> * lines 28f: "three retrieval papers: Torricella.., McGuire ..., and Teng ... .
> * line 31: "Legendre series, respectively:"
> * equation after line 31: please introduce g_l, P_l
> * line 33: "solving the equation"
> * equation after line 33: what is sub-/superscript m, Q^m
> * line 35: "using a quadrature scheme"
> * line 38: I would move the two references to the end of sentence
> * line 42: formatting of delta-m (should be delta-_M_), could probably use another reference to Wiscombe, 1977
> * lines 43 and 45: it does not seem necessary to introduce acronyms here since they are not reused unless they are instructional for readers
> * line 49: "PhythonicDISORT is the first reimplementation of DISORT from scratch in Python"
> * line 52: "remans slower than"
> * line 60: please check meaning "upscale computational performance using DISORT"
> * line 64: "system-dependent"
> * line 65: "shell scripts" (plural?, or "a shell script")
> * line 67: "widely used"
> * lines 67f: suggest: "should be more easily understandable for Python users than DISORT's FORTRAN implementation"
> * line 147: formatting of delta-m (should be delta-_M_)

@phibeck Thank you very much! Noted and I will make the changes soon.
