+++ docs/JOSS_Paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Wis1977,
       author = "W. J.  Wiscombe",
-      title = "The Delta–M Method: Rapid Yet Accurate Radiative Flux Calculations for Strongly Asymmetric Phase Functions",
+      title = "The Delta–$M$ Method: Rapid Yet Accurate Radiative Flux Calculations for Strongly Asymmetric Phase Functions",
       journal = "Journal of Atmospheric Sciences",
       year = "1977",
       publisher = "American Meteorological Society",
