+++ docs/JOSS_Paper/paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0009-0000-5829-5081
     affiliation: "1"
 affiliations:
- - name: Columbia University, Department of Applied Physics and Applied Mathematics, USA
+ - name: Columbia University, Department of Applied Physics and Applied Mathematics, United States of America
    index: 1
 date: 11 February 2024
 bibliography: paper.bib
@@ -71,13 +71,12 @@ Dirichlet boundary conditions.
 The Radiative Transfer Equation (RTE) models the processes of absorption, scattering and emission 
 as electromagnetic radiation propagates through a medium. 
 Consider a plane-parallel, horizontally homogeneous atmosphere with vertical coordinate 
-$\tau$ (optical depth) increasing from top to bottom and directional coordinates $\phi$ for the azimuthal angle (positive is counterclockwise) 
-and $\mu=\cos\theta$ for the polar direction ($\theta$ is the polar angle measured from the surface normal), 
-with $\mu > 0$ pointing up following the convention of [@STWJ1988].
-Given three possible sources: 
-blackbody emission from the atmosphere $s(\tau)$; 
-scattering from a collimated beam of starlight with intensity $I_0$ and incident azimuthal and cosine polar angles $\phi_0, \mu_0$;
-radiation from other atmospheric layers or the Earth's surface which is modeled by Dirichlet boundary conditions,
+$\tau$ (optical depth) increasing from top to bottom, directional coordinates $\phi$ for the azimuthal angle (positive is counterclockwise), 
+and $\mu=\cos\theta$ for the polar direction ($\theta$ is the polar angle measured from the surface normal) 
+with $\mu > 0$ pointing up following the convention of @STWJ1988.
+Given three possible sources, namely blackbody emission from the atmosphere $s(\tau)$, 
+scattering from a collimated beam of starlight with intensity $I_0$ as well as incident azimuthal and cosine polar angles \phi_0 and \mu_0, respectively,
+and radiation from other atmospheric layers or the Earth's surface which is modeled by Dirichlet boundary conditions,
 the diffuse intensity $u(\tau, \mu, \phi)$ propagating in direction $(\mu, \phi)$ 
 is described by the 1D RTE [@Cha1960; @STWJ1988]:
 
@@ -96,12 +95,12 @@ a multi-layer atmosphere with different $\omega$ and $p$ for each layer.
 The RTE is important in many fields of science and engineering,
 for example, in the retrieval of optical properties of the medium from measurements [@TCCGL1999; @MRO/CRISM2008; @TLZWSY2020].
 The gold standard for numerically solving the 1D RTE is the Discrete Ordinate Radiative Transfer 
-package `DISORT` which was coded in FORTRAN 77 and first released in 1988 [@STWJ1988; @Sta1999].
+package `DISORT` which was coded in FORTRAN77 and first released in 1988 [@STWJ1988; @Sta1999].
 It has been widely used, for example by `MODTRAN` [@Ber2014], `Streamer` [@Key1998], and `SBDART` [@Ric1998],
 all of which are comprehensive radiative transfer models that are themselves widely used in atmospheric science,
-and by the three retrieval papers @TCCGL1999; @MRO/CRISM2008; @TLZWSY2020.
+and by the three retrieval papers: @TCCGL1999; @MRO/CRISM2008; @TLZWSY2020.
 `DISORT` implements the Discrete Ordinates Method which has two key steps.
-First, the diffuse intensity function $u$ and phase function $p$ are expanded as the Fourier cosine series and Legendre series respectively:
+First, the diffuse intensity function $u$ and phase function $p$ are expanded as the Fourier cosine series and Legendre series, respectively:
 
 $$
[…]
