### 2024-11-13 c30b95d
Continued edits to JOSS paper
