### 2024-11-13 #6442 @dhjx1996
Hi @phibeck I have made the changes. Two clarifications:

> * line 42: formatting of delta-m (should be delta-_M_), could probably use another reference to Wiscombe, 1977

I reformatted delta-_M_ but did not add the reference. Most entries in that list / paragraph have associated references which have been cited elsewhere but I think adding them would bog down what is otherwise intended to be succinct list of features.

> * lines 67f: suggest: "should be more easily understandable for Python users than DISORT's FORTRAN implementation"

I changed the sentence to read: "Given that Python is a widely used high-level language, `PythonicDISORT`'s
code should be accessible to more people than `DISORT`'s FORTRAN code."
