### 2024-11-15 #6442 @phibeck
Hi @dhjx1996 thanks for the update. Just two more small edits, then this is good to go from my side. Thanks!
- line 26: in my previous note I was referring to changing "coded" to "written", you can change FORTRAN77 back to FORTRAN 77 (this was incorrect in my suggestion)
- line 30: in your list of retrieval papers, please change semicolons to comma and add an "and", i.e.: "three retrieval papers: Torricella et al. (1999), McGuire et al. (2008), and Teng et al. (2020).
