### 2024-11-18 #6442 @kyleniemeyer
@dhjx1996 could you remove the code modifiers in the paper title? That is leading to some `<code>` fields on the paper's web page.

Otherwise, congratulations on your article's publication in JOSS! Please consider [signing up as a reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.
