### 2024-11-18 38e4239
Removed code from title of JOSS paper
