### 2024-11-18 #6442 @dhjx1996
Hi @kyleniemeyer, the code has been removed. Thank you @phibeck, @arjunsavel and @simonrp84!
